update LogoSuite stories
