+++ packages/react/src/LogoSuite/LogoSuite.fixtures.tsx
@@ -1,0 +1,164 @@
+import React from 'react'
+
+export const logos = [
+  <svg key={1} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 115 43" fill="none" focusable="false">
+    <title>Uber</title>
+    <g>
+      <path
+        fill="#24292F"
+        fillRule="evenodd"
+        d="M8.743 33.353c-.847-.902-1.51-1.989-1.99-3.26-.478-1.27-.718-2.68-.718-4.227V1.996H.233v24.202c0 2.284.378 4.356 1.133 6.216.755 1.86 1.787 3.453 3.095 4.78a13.519 13.519 0 0 0 4.641 3.066c1.787.719 3.711 1.078 5.774 1.078 2.027 0 3.914-.368 5.664-1.105a13.322 13.322 0 0 0 4.56-3.15v3.592h5.746V1.995h-5.802v23.871c0 1.547-.24 2.947-.718 4.2-.48 1.252-1.152 2.33-2.018 3.232a9.176 9.176 0 0 1-3.038 2.1 9.368 9.368 0 0 1-3.73.746c-1.363 0-2.625-.24-3.786-.718a8.56 8.56 0 0 1-3.011-2.073Zm32.104-17.267V1.996h-5.581v38.679h5.526v-3.537a14.678 14.678 0 0 0 4.503 3.067c1.713.755 3.564 1.133 5.553 1.133 2.063 0 3.997-.387 5.802-1.16 1.805-.774 3.37-1.824 4.697-3.15a14.906 14.906 0 0 0 3.15-4.67c.773-1.786 1.16-3.692 1.16-5.718s-.387-3.942-1.16-5.747c-.774-1.805-1.824-3.37-3.15-4.697-1.326-1.326-2.892-2.376-4.697-3.15-1.805-.773-3.739-1.16-5.802-1.16-1.989 0-3.83.378-5.525 1.133a14.78 14.78 0 0 0-4.476 3.067Zm2.707 17.46a9.611 9.611 0 0 1-2.072-3.094c-.497-1.178-.746-2.45-.746-3.812 0-1.363.258-2.643.774-3.84a9.913 9.913 0 0 1 2.1-3.123 9.621 9.621 0 0 1 3.094-2.072 9.567 9.567 0 0 1 3.758-.746 9.57 9.57 0 0 1 3.757.746 9.303 9.303 0 0 1 3.067 2.072 10.154 10.154 0 0 1 2.072 3.122 9.603 9.603 0 0 1 .773 3.84c0 1.364-.258 2.635-.773 3.813a10.247 10.247 0 0 1-2.072 3.095 9.666 9.666 0 0 1-3.067 2.1 9.272 9.272 0 0 1-3.758.773c-1.362 0-2.633-.258-3.812-.773a9.999 9.999 0 0 1-3.095-2.1Zm25.252-12.598a15.03 15.03 0 0 1 3.04-4.669 14.12 14.12 0 0 1 4.586-3.177c1.768-.774 3.683-1.16 5.746-1.16 2.026 0 3.905.368 5.637 1.105 1.73.736 3.223 1.759 4.475 3.066 1.253 1.308 2.23 2.855 2.929 4.642.7 1.787 1.05 3.73 1.05 5.83v1.823H73.337a9.58 9.58 0 0 0 3.177 5.719 9.47 9.47 0 0 0 2.819 1.685 9.204 9.204 0 0 0 3.343.608c3.352 0 6.06-1.381 8.123-4.144l4.033 2.984a15.007 15.007 0 0 1-5.194 4.42c-2.063 1.069-4.384 1.603-6.962 1.603-2.1 0-4.061-.369-5.885-1.105-1.823-.737-3.407-1.769-4.752-3.095a14.68 14.68 0 0 1-3.177-4.669c-.774-1.787-1.16-3.73-1.16-5.83-.001-1.989.368-3.867 1.104-5.636Zm7.764-2.182c-1.566 1.307-2.606 3.048-3.122 5.221h17.185c-.48-2.173-1.501-3.914-3.067-5.221-1.566-1.308-3.398-1.962-5.498-1.962s-3.933.654-5.498 1.962Zm28.374 5.884c0-2.21.589-3.96 1.768-5.25 1.179-1.288 2.689-1.933 4.531-1.933h2.321v-5.194h-1.934c-1.547 0-2.883.331-4.006.994a7.853 7.853 0 0 0-2.736 2.708v-3.481h-5.525v28.18h5.581V24.65Z"
+        clipRule="evenodd"
+      />
+    </g>
+    <defs>
+      <clipPath>
+        <path fill="#fff" d="M.233.452h114.75v42.43H.233z" />
+      </clipPath>
+    </defs>
+  </svg>,
+  <svg key={2} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 165 41" fill="none">
+    <title>Vercel</title>
+    <path
+      fill="#24292F"
+      d="M81.939 11.46c-6.33 0-10.9 4.155-10.9 10.374 0 6.22 5.138 10.376 11.468 10.376 3.84 0 7.194-1.521 9.304-4.074l-4.382-2.553c-1.163 1.276-2.92 2.01-4.895 2.01-2.76 0-5.085-1.44-5.95-3.749h16.093a9.74 9.74 0 0 0 .19-2.01c-.028-6.246-4.572-10.375-10.928-10.375Zm-5.437 8.337c.73-2.308 2.678-3.748 5.437-3.748 2.759 0 4.706 1.44 5.436 3.748H76.502Zm67.242-8.338c-6.329 0-10.901 4.156-10.901 10.375 0 6.22 5.14 10.376 11.469 10.376 3.841 0 7.195-1.521 9.304-4.074l-4.381-2.553c-1.163 1.276-2.922 2.01-4.896 2.01-2.759 0-5.085-1.44-5.951-3.749h16.094a9.762 9.762 0 0 0 .189-2.01c0-6.246-4.571-10.375-10.927-10.375Zm-5.41 8.338c.731-2.308 2.678-3.748 5.437-3.748 2.759 0 4.706 1.44 5.437 3.748h-10.874Zm-22.396 2.037c0 3.45 2.245 5.759 5.735 5.759 2.353 0 4.138-1.087 5.058-2.825l4.408 2.553c-1.812 3.042-5.247 4.889-9.466 4.889-6.357 0-10.901-4.156-10.901-10.376s4.571-10.375 10.901-10.375c4.219 0 7.627 1.847 9.466 4.89l-4.408 2.552c-.92-1.765-2.678-2.825-5.058-2.825-3.49-.027-5.735 2.282-5.735 5.758Zm47.335-16.73v26.508h-5.167V5.104h5.167ZM22.189 2.224 43.395 39.11H.983L22.19 2.224Zm53.042 2.88L59.326 32.78 43.422 5.104H49.4l9.953 17.3 9.954-17.3h5.924Zm33.81 6.926v5.595a6.684 6.684 0 0 0-1.839-.272c-3.327 0-5.734 2.309-5.734 5.758v8.528h-5.167V12.03h5.167v5.297c0-2.934 3.381-5.296 7.573-5.296Z"
+    />
+  </svg>,
+  <svg key={3} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 174 50" width="174" height="50" fill="none">
+    <title>Shopify</title>
+    <path
+      fill="#24292F"
+      d="M60.128 27.374c-1.469-.812-2.242-1.469-2.242-2.397 0-1.198 1.044-1.933 2.706-1.933 1.933 0 3.634.812 3.634.812l1.353-4.137s-1.237-.966-4.91-.966c-5.103 0-8.66 2.938-8.66 7.036 0 2.32 1.663 4.098 3.866 5.374 1.779 1.005 2.397 1.74 2.397 2.783 0 1.122-.889 2.01-2.551 2.01-2.474 0-4.794-1.275-4.794-1.275l-1.43 4.137s2.164 1.43 5.76 1.43c5.258 0 9.008-2.59 9.008-7.23.077-2.551-1.856-4.33-4.137-5.644ZM81.082 18.676c-2.59 0-4.6 1.237-6.186 3.093l-.077-.04 2.242-11.752h-5.876l-5.683 29.885h5.838l1.933-10.207c.773-3.866 2.744-6.224 4.6-6.224 1.315 0 1.817.889 1.817 2.165 0 .812-.077 1.778-.27 2.59L77.216 39.9h5.838l2.28-12.062c.271-1.276.426-2.783.426-3.827.077-3.364-1.624-5.335-4.678-5.335ZM99.06 18.676c-7.037 0-11.715 6.34-11.715 13.415 0 4.523 2.784 8.196 8.041 8.196 6.921 0 11.56-6.186 11.56-13.415.038-4.214-2.397-8.196-7.887-8.196Zm-2.862 17.088c-2.01 0-2.822-1.701-2.822-3.828 0-3.363 1.74-8.814 4.91-8.814 2.088 0 2.745 1.778 2.745 3.518 0 3.634-1.74 9.124-4.833 9.124ZM121.946 18.676c-3.943 0-6.186 3.48-6.186 3.48h-.077l.348-3.132h-5.181c-.27 2.126-.734 5.335-1.198 7.77l-4.059 21.341h5.837l1.624-8.66h.116s1.199.773 3.441.773c6.881 0 11.366-7.036 11.366-14.15 0-3.904-1.74-7.422-6.031-7.422Zm-5.567 17.204c-1.508 0-2.397-.85-2.397-.85l.966-5.452c.696-3.634 2.591-6.07 4.601-6.07 1.778 0 2.32 1.663 2.32 3.21 0 3.75-2.243 9.162-5.49 9.162ZM136.328 10.286c-1.856 0-3.364 1.47-3.364 3.402 0 1.74 1.083 2.938 2.745 2.938h.077c1.817 0 3.402-1.237 3.441-3.402.039-1.74-1.121-2.938-2.899-2.938ZM128.132 39.862h5.876l3.982-20.722h-5.915l-3.943 20.722ZM152.836 19.1h-4.06l.194-.966c.347-2.01 1.507-3.788 3.479-3.788 1.044 0 1.856.309 1.856.309l1.159-4.562s-1.005-.503-3.17-.503c-2.087 0-4.136.58-5.721 1.933-2.011 1.701-2.939 4.137-3.402 6.611l-.155.967h-2.706l-.851 4.407h2.706l-3.092 16.353h5.837l3.093-16.353h4.021l.812-4.407ZM166.908 19.14s-3.673 9.2-5.297 14.227h-.077c-.116-1.624-1.43-14.227-1.43-14.227h-6.147l3.518 19.02c.077.426.038.696-.116.967-.696 1.314-1.817 2.59-3.17 3.518-1.083.812-2.32 1.315-3.286 1.662l1.623 4.949c1.199-.27 3.634-1.237 5.722-3.17 2.668-2.513 5.142-6.34 7.655-11.598l7.152-15.348h-6.147ZM29.354 5.686s-.54.154-1.43.425a11.31 11.31 0 0 0-.696-1.701c-1.005-1.933-2.513-2.977-4.291-2.977-.116 0-.232 0-.387.039-.038-.078-.116-.116-.154-.194C21.622.428 20.617.041 19.419.08c-2.32.077-4.64 1.74-6.495 4.716-1.315 2.088-2.32 4.717-2.59 6.766-2.668.812-4.524 1.392-4.562 1.43-1.353.426-1.392.464-1.547 1.74C4.11 15.699.552 42.915.552 42.915L29.78 47.98V5.647c-.232 0-.348.039-.426.039ZM22.59 7.773c-1.546.464-3.248 1.005-4.91 1.508.464-1.817 1.392-3.634 2.474-4.833.425-.425 1.005-.927 1.663-1.237.657 1.392.812 3.286.773 4.562Zm-3.17-6.108c.541 0 1.005.116 1.392.348-.62.31-1.238.812-1.817 1.392-1.47 1.585-2.59 4.059-3.055 6.417-1.392.425-2.783.85-4.059 1.237.85-3.711 3.982-9.317 7.539-9.394Zm-4.485 21.263c.155 2.474 6.688 3.015 7.075 8.853.27 4.6-2.436 7.732-6.34 7.964-4.717.31-7.307-2.474-7.307-2.474l1.005-4.253s2.59 1.972 4.678 1.817c1.353-.077 1.856-1.198 1.817-1.971-.193-3.248-5.528-3.054-5.876-8.39-.31-4.484 2.628-9.007 9.123-9.433 2.513-.154 3.79.464 3.79.464l-1.47 5.567s-1.662-.773-3.634-.618c-2.86.193-2.9 2.01-2.86 2.474Zm9.201-15.619c0-1.16-.154-2.822-.696-4.214 1.779.348 2.63 2.32 3.016 3.518-.696.194-1.47.426-2.32.696ZM30.669 47.903l12.14-3.016S37.588 9.59 37.55 9.358a.45.45 0 0 0-.426-.386c-.193 0-3.595-.078-3.595-.078s-2.088-2.01-2.861-2.783v41.792Z"
+    />
+  </svg>,
+  <svg key={4} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 175 45" fill="none">
+    <title>Pinterest</title>
+    <path
+      fill="#24292F"
+      fillRule="evenodd"
+      d="M.088 16.766c0-8.45 6.12-16.23 17.716-16.23 9.295 0 16.521 6.644 16.55 15.53 0 9.267-5.827 16.697-13.927 16.697-2.71 0-5.274-1.428-6.149-3.089 0 0-1.34 5.129-1.66 6.381-.904 3.468-4.109 8.276-4.371 8.625a.245.245 0 0 1-.204.117c-.146 0-.292-.117-.292-.262-.087-.496-1.02-6.265-.204-9.82.496-2.098 3.147-13.345 3.147-13.345a9.424 9.424 0 0 1-.815-3.992c0-3.73 2.156-6.527 4.866-6.527 2.272 0 3.38 1.72 3.38 3.788 0 1.424-.558 3.284-1.15 5.254-.363 1.213-.74 2.47-1.007 3.691-.64 2.652 1.34 4.837 3.963 4.837 4.75 0 8.421-5.011 8.421-12.267 0-6.41-4.604-10.898-11.189-10.898-7.634 0-12.121 5.712-12.121 11.627 0 2.302.874 4.778 1.981 6.119.175.233.262.524.175.786-.106.44-.283 1.137-.44 1.76-.146.577-.276 1.089-.318 1.27-.116.496-.379.583-.903.35C2.187 25.595.088 20.7.088 16.766Zm91.377 4.837c1.865 0 2.506 1.137 2.506 2.477 0 2.273-1.748 3.642-4.517 3.642a12.9 12.9 0 0 1-3.176-.495c.35-2.098 1.69-5.624 5.187-5.624Zm39.773 0c1.69 0 2.506 1.137 2.506 2.477 0 2.273-1.748 3.642-4.516 3.642a12.901 12.901 0 0 1-3.176-.495c.35-2.098 1.865-5.624 5.186-5.624Zm37.705.175h4.225c.583 0 .933-.233 1.049-.787.074-.286.164-.65.254-1.012v-.002c.157-.635.313-1.27.387-1.492.117-.495-.146-.845-.699-.845h-4.196c.051-.207.128-.496.222-.845l.001-.002c.448-1.663 1.263-4.692 1.263-6.787 0-5.07-4.312-9.382-10.431-9.382-6.906 0-11.393 4.633-11.393 11.13a10.809 10.809 0 0 0 1.602 5.508c-.379-.03-.757-.03-1.136-.03-5.245 0-8.538 2.856-8.538 6.994 0 3.875 3.06 5.39 5.944 6.41.393.138.78.265 1.155.387 2.077.679 3.77 1.232 3.77 2.935 0 1.34-1.282 2.127-3.846 2.127-3.002 0-4.662-1.428-5.1-1.953a2.08 2.08 0 0 0 .7-1.66c0-1.253-.874-2.244-2.506-2.244-1.37 0-2.535 1.107-2.827 2.593-2.214 1.807-4.603 2.972-8.012 2.972-3.38 0-5.1-1.748-5.1-5.477.787.233 2.856.524 4.196.524 5.682 0 9.558-2.593 9.558-7.78 0-2.797-2.448-5.711-7.285-5.711h-.029c-6.352 0-9.936 4.021-11.481 8.48-.218-.06-.437-.11-.655-.161-.219-.051-.437-.102-.656-.16a9.626 9.626 0 0 0 .612-3.38c0-2.186-1.369-4.779-4.982-4.779-2.535 0-4.925 1.66-6.586 4.138.186-.742.347-1.364.468-1.83.145-.558.232-.89.232-.939.116-.437-.059-.903-.729-.903h-3.729c-.525 0-.904.175-1.049.787-.088.408-1.37 5.449-2.623 10.315-.787 3.06-4.837 6.847-9.936 6.847-3.35 0-5.1-1.865-5.1-5.594.875.35 2.827.64 4.197.64 5.915 0 9.586-2.592 9.586-7.78 0-2.796-2.447-5.71-7.284-5.71h-.03c-7.167 0-11.538 5.565-12.296 11.888-.32 2.623-2.535 6.236-4.924 6.236-1.166 0-1.836-.729-1.836-1.982 0-.851.473-2.665 1.087-5.027.164-.629.338-1.296.516-1.995.188-.715.425-1.673.694-2.764l.002-.006c.147-.597.305-1.234.47-1.892h3.904c.583 0 .903-.233 1.049-.787.175-.786.524-2.156.64-2.506.117-.495-.145-.845-.698-.845h-3.847s1.749-7.022 1.807-7.226c.175-.728-.437-1.136-1.05-.99 0 0-2.971.582-3.525.699-.553.116-.99.437-1.194 1.224l-1.574 6.322h-3.06c-.582 0-.932.234-1.048.787-.075.286-.164.65-.254 1.013-.157.635-.313 1.27-.387 1.493-.117.495.145.845.699.845h3.001c-.029.088-1.078 4.138-1.952 7.78-.408 1.748-1.778 5.857-3.992 5.857-1.34 0-1.894-.67-1.894-2.127 0-1.102.528-3.006 1.071-4.967.577-2.084 1.173-4.231 1.173-5.552 0-3.438-1.836-5.449-5.624-5.449-2.39 0-4.866 1.544-5.915 2.885 0 0 .32-1.108.437-1.545a.676.676 0 0 0-.437-.845c-.087-.029-.204-.029-.291-.029h-3.672c-.787 0-.99.437-1.107.874-.032.095-.429 1.669-.99 3.89-.477 1.892-1.072 4.253-1.662 6.57-.874 3.468-3 6.353-5.274 6.353-1.165 0-1.69-.728-1.69-1.981 0-.945.537-3.098 1.22-5.831.095-.387.195-.784.296-1.192a1023.18 1023.18 0 0 1 1.396-5.433c.326-1.256.544-2.095.585-2.288.117-.495-.087-.904-.729-.904h-3.73c-.67 0-.932.35-1.048.845 0 0-1.05 3.963-2.156 8.392l-.223.873c-.745 2.91-1.467 5.736-1.467 7.14 0 2.74 1.223 4.808 4.603 4.808 2.594 0 4.663-1.311 6.236-3.001l-.14.546c-.151.587-.245.953-.268.998-.146.525.03.99.641.99h3.817c.67 0 .903-.261 1.05-.844.145-.583 2.971-11.743 2.971-11.743.758-3.001 2.623-5.012 5.245-5.012 1.253 0 2.302.816 2.185 2.419-.057.691-.42 2.097-.838 3.716-.638 2.476-1.405 5.45-1.405 7.123 0 2.098.786 4.808 4.72 4.808 2.68 0 4.691-1.282 6.12-2.972.524 1.777 1.922 2.972 4.486 2.972 4.255 0 6.819-2.535 8.334-5.07 1.253 2.943 4.196 5.07 8.48 5.07a13.727 13.727 0 0 0 10.168-4.37c-.208.803-.382 1.488-.512 1.995-.143.559-.23.901-.245.947-.146.525.087.962.728.962h3.759a.981.981 0 0 0 1.02-.816c.025-.112.141-.562.325-1.275l.001-.004v-.002c.245-.95.611-2.369 1.043-4.08 1.486-5.711 3.846-11.656 7.401-11.656 1.253 0 1.778.7 1.778 1.807 0 .408-.088.845-.292 1.224-1.719-.35-3.117.495-3.117 2.33 0 1.195 1.282 2.303 2.972 2.303 1.311 0 2.389-.32 3.234-.904.306.069.626.137.954.206h.001c.513.11 1.047.224 1.58.348a17.73 17.73 0 0 0-.204 2.506c0 4.604 3.438 8.508 9.324 8.508 4.604 0 7.576-1.544 10.344-3.846 1.311 2.156 4.633 3.875 8.975 3.875 5.944 0 9.353-3.088 9.353-7.197 0-3.7-3.059-5.07-6.177-6.206-.311-.113-.616-.22-.911-.323-2.138-.75-3.78-1.324-3.78-2.912 0-1.486 1.165-2.156 3.147-2.156.932 0 1.893.146 2.797.437.524 1.282 1.573 2.71 3.38 2.71 1.544 0 2.214-1.224 2.214-2.33 0-1.288-.87-2.11-1.906-3.09-1.52-1.437-3.397-3.21-3.397-7.284 0-4.254 2.419-7.75 7.401-7.75 3.468 0 5.74 2.185 5.74 5.74 0 2.913-1.777 8.129-1.777 8.129h-3.001c-.583 0-.933.233-1.049.787-.074.286-.164.65-.254 1.012-.156.636-.313 1.271-.387 1.494-.117.495.145.845.699.845h2.943s-3.03 11.218-3.03 13.257c0 2.74 1.544 4.808 4.924 4.808 4.779 0 7.78-3.089 9.062-7.605a.484.484 0 0 0-.233-.612c-.058-.029-.146-.029-.204-.029h-1.836a.686.686 0 0 0-.67.554c-.408 1.602-1.428 3.292-3.438 3.292-1.166 0-1.836-.728-1.836-1.981 0-.953.546-3.021 1.297-5.863l.306-1.16c.437-1.748 1.165-4.778 1.165-4.778Zm-124.594-9.85c-.146 1.982-1.777 3.556-3.671 3.556s-3.322-1.603-3.206-3.555c.146-1.982 1.778-3.555 3.672-3.555 1.894 0 3.322 1.603 3.205 3.555Z"
+      clipRule="evenodd"
+    />
+  </svg>,
+  <svg key={5} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 134 40" width="134" height="40" fill="none">
+    <title>Twilio</title>
+    <path
+      fill="#57606A"
+      fillRule="evenodd"
+      d="M93.765 7.84h-7.04c-.16 0-.32.16-.32.32v4.08c0 .16.16.32.32.32h7.04c.16 0 .32-.16.32-.32V8.16c.08-.16-.08-.32-.32-.32Zm-.08 6.08h-13.6c-.16 0-.4.16-.48.32l-1.76 6.64-.08.32-2.08-7.12c-.08-.16-.24-.32-.48-.32h-5.28c-.16 0-.4.16-.48.32l-2 6.64-.08.32-.08-.16-.8-3.28-.88-3.28c-.08-.16-.24-.32-.48-.32h-10.64V8.24c0-.16-.24-.4-.48-.32L47.285 10c-.24.08-.32.16-.32.32v3.6h-1.76c-.16 0-.32.16-.32.32v5.12c0 .16.16.32.32.32h1.76v6.24c0 4.4 2.4 6.48 6.8 6.48 1.84 0 3.6-.4 4.8-1.04v-5.28c0-.24-.24-.32-.4-.32-.72.32-1.28.4-1.92.4-1.2 0-1.84-.48-1.84-1.84v-4.56h3.84c.16 0 .32-.16.32-.32v-4.32l5.04 16.72c.08.16.24.32.48.32h5.6c.16 0 .4-.16.48-.32l2.4-7.44 1.2 3.84 1.12 3.52c.08.16.24.32.48.32h5.6c.16 0 .4-.16.48-.32l5.12-16.8v16.8c0 .16.16.32.32.32h6.8c.16 0 .32-.16.32-.32V14.32c0-.24-.16-.4-.32-.4Zm8.96-6.08h-6.8c-.16 0-.32.16-.32.32v23.6c0 .16.16.32.32.32h6.8c.16 0 .32-.16.32-.32V8.16c0-.16-.16-.32-.32-.32Zm9.04 0h-7.04c-.16 0-.32.16-.32.32v4.08c0 .16.16.32.32.32h7.04c.16 0 .32-.16.32-.32V8.16c.08-.16-.08-.32-.32-.32Zm-.08 6.08h-6.8c-.16 0-.32.16-.32.32v17.52c0 .16.16.32.32.32h6.8c.16 0 .32-.16.32-.32V14.32c0-.24-.16-.4-.32-.4Zm11.28-.4c-6 0-10.24 4.4-10.24 9.52v.08c0 5.12 4.24 9.44 10.16 9.44 6 0 10.24-4.4 10.24-9.52v-.08c0-5.12-4.24-9.44-10.16-9.44Zm2.88 9.52c0 1.76-1.2 3.12-2.88 3.12s-2.96-1.44-2.96-3.2v-.08c0-1.76 1.2-3.12 2.88-3.12s2.96 1.44 2.96 3.28ZM20.085 0c-11.04 0-20 8.96-20 20s8.96 20 20 20 20-8.96 20-20-8.96-20-20-20Zm0 34.72c-8.16 0-14.72-6.56-14.72-14.72s6.56-14.72 14.72-14.72 14.72 6.56 14.72 14.72-6.56 14.72-14.72 14.72Zm4.96-15.52a4.16 4.16 0 1 0 0-8.32 4.16 4.16 0 0 0 0 8.32Zm4.16 5.76a4.16 4.16 0 1 1-8.32 0 4.16 4.16 0 0 1 8.32 0Zm-14.08 4.16a4.16 4.16 0 1 0 0-8.32 4.16 4.16 0 0 0 0 8.32Zm4.16-14.08a4.16 4.16 0 1 1-8.32 0 4.16 4.16 0 0 1 8.32 0Z"
+      clipRule="evenodd"
+    />
+  </svg>,
+  <svg key={6} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 156 41" width="156" height="41" fill="none">
+    <title>Slack</title>
+    <path
+      fill="#24292F"
+      fillRule="evenodd"
+      d="M14.558 9.265a4.13 4.13 0 0 1-4.135-4.135A4.13 4.13 0 0 1 14.558.995a4.13 4.13 0 0 1 4.134 4.135v4.135h-4.134ZM71.505 1.59h6.703v32.796h-6.703V1.59ZM49.797 31.975l1.942-4.511c2.099 1.566 4.887 2.38 7.643 2.38 2.036 0 3.32-.783 3.32-1.973-.011-1.285-1.84-1.683-4.081-2.172-3.55-.773-8.14-1.773-8.197-6.88-.032-4.23 3.727-7.487 9.052-7.487 3.164 0 6.328.783 8.583 2.569l-1.817 4.604c-2.067-1.315-4.636-2.255-7.079-2.255-1.66 0-2.756.783-2.756 1.785.012 1.272 1.886 1.775 4.174 2.39 3.572.959 8.153 2.189 8.23 7.039 0 4.323-3.665 7.361-8.928 7.361-3.852 0-7.392-.908-10.086-2.85Zm74.52-6.14c-.971 1.692-2.788 2.85-4.887 2.85a5.601 5.601 0 0 1-5.607-5.606 5.601 5.601 0 0 1 5.607-5.607c2.099 0 3.916 1.159 4.887 2.85l5.356-2.976c-2.004-3.57-5.857-6.014-10.243-6.014-6.484 0-11.746 5.262-11.746 11.747 0 6.484 5.262 11.746 11.746 11.746 4.417 0 8.239-2.412 10.243-6.014l-5.356-2.976Zm7.987-24.245v32.796h6.704v-9.835l7.956 9.835h8.583l-10.118-11.683 9.366-10.901h-8.207l-7.58 9.053V1.59h-6.704ZM92.899 28.686c2.255 0 4.26-1.19 5.231-2.788v-5.545c-.971-1.628-2.976-2.881-5.231-2.881a5.601 5.601 0 0 0-5.607 5.607 5.601 5.601 0 0 0 5.607 5.607Zm5.231-14.222v-2.662h6.703v22.553H98.13v-2.662c-1.096 1.848-3.822 3.132-6.672 3.132-5.889 0-10.525-5.262-10.525-11.778 0-6.515 4.636-11.715 10.525-11.715 2.85 0 5.576 1.284 6.672 3.132ZM8.356 25.866a4.13 4.13 0 0 1-4.135 4.135 4.13 4.13 0 0 1-4.135-4.135 4.13 4.13 0 0 1 4.135-4.134h4.135v4.134Zm6.202-4.134a4.13 4.13 0 0 0-4.135 4.134v10.337a4.13 4.13 0 0 0 4.135 4.135 4.13 4.13 0 0 0 4.135-4.135V25.866a4.13 4.13 0 0 0-4.135-4.134Zm4.134-6.234a4.13 4.13 0 0 0-4.134-4.135H4.19a4.13 4.13 0 0 0-4.135 4.135 4.13 4.13 0 0 0 4.135 4.135h10.368a4.13 4.13 0 0 0 4.134-4.135Zm12.436 0a4.13 4.13 0 0 1 4.135-4.135 4.13 4.13 0 0 1 4.135 4.135 4.13 4.13 0 0 1-4.135 4.135h-4.135v-4.135Zm-6.202 4.135a4.13 4.13 0 0 0 4.135-4.135V5.13A4.13 4.13 0 0 0 24.926.995a4.13 4.13 0 0 0-4.135 4.135v10.368a4.13 4.13 0 0 0 4.135 4.135Zm0 12.436a4.13 4.13 0 0 1 4.135 4.134 4.13 4.13 0 0 1-4.135 4.135 4.13 4.13 0 0 1-4.135-4.135V32.07h4.135Zm-4.135-6.203a4.13 4.13 0 0 0 4.135 4.135h10.368a4.13 4.13 0 0 0 4.135-4.135 4.13 4.13 0 0 0-4.135-4.134H24.926a4.13 4.13 0 0 0-4.135 4.134Z"
+      clipRule="evenodd"
+    />
+  </svg>,
+  <svg key={7} role="img" width="38" height="48" viewBox="0 0 38 48" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <title>Lion</title>
+    <path
+      d="M17.7693 8.28307L18.7622 9.11922L19.7551 8.25694L19.729 3.37071H17.7431L17.7693 8.28307ZM18.7361 2.5607C19.4416 2.5607 20.0164 1.98585 20.0164 1.28035C20.0164 0.57485 19.4416 0 18.7099 0C18.0044 0 17.4296 0.57485 17.4296 1.28035C17.4557 1.98585 18.0306 2.5607 18.7361 2.5607ZM9.46009 9.24987H12.0469C12.0469 9.24987 11.6288 8.83179 10.0349 7.13337C9.12041 6.16658 9.04202 5.40882 10.1917 5.40882H10.453L11.7595 3.4491C11.7595 3.4491 10.6359 3.42297 10.1656 3.42297C7.08231 3.42297 6.53359 6.454 8.07523 7.91726C9.25106 8.98857 9.46009 9.24987 9.46009 9.24987ZM27.3065 5.33043C28.4562 5.33043 28.3779 6.08819 27.4633 7.05498L25.4775 9.17148H28.0643C28.0643 9.17148 28.2995 8.93631 29.4492 7.81274C30.9908 6.34948 30.416 3.31845 27.3588 3.31845C26.8623 3.31845 25.7649 3.37071 25.7649 3.37071L27.0714 5.33043H27.3065ZM34.3615 29.4219C34.858 31.7736 34.7012 34.1252 33.8912 36.4246C32.8721 39.2466 31.3044 41.1541 29.3447 41.8857C28.9004 42.0425 28.0904 42.3038 27.0714 42.3299C26.4181 42.3299 25.5559 42.0947 25.0071 41.8334C24.6936 41.7028 24.0404 41.3892 24.0404 41.3892C22.2897 40.4224 19.8074 38.4627 19.8335 37.1562C19.8335 36.712 20.8526 35.9804 21.7148 35.9543C22.9168 35.9543 23.8836 34.9614 23.8836 33.7594V33.6549L23.4394 31.7474L23.4132 31.6952C23.0997 30.65 22.2897 27.0964 25.0594 25.5025L27.4633 24.3528L26.6011 22.5498L24.1449 23.7256L24.0665 23.7518C21.3229 25.3196 20.3561 28.4812 21.4535 32.2439L21.8193 33.8639C21.7932 33.9162 21.741 33.9684 21.6626 33.9684C20.8526 33.9684 19.938 34.282 19.2064 34.8046C19.0496 34.9091 18.919 35.0136 18.7883 35.1443C17.9783 34.4126 16.8025 33.9946 15.888 34.0207C15.8096 34.0207 15.7573 33.9684 15.7312 33.9162L16.097 32.2961C17.1683 28.5335 16.1754 25.3718 13.4056 23.8302L10.8972 22.6543L10.0611 24.4573L12.465 25.5808C15.2347 27.1486 14.477 30.7022 14.1634 31.7736L13.7192 33.7333V33.8378C13.7192 35.0397 14.7121 36.0065 15.9141 36.0065C16.7502 36.0065 17.7954 36.712 17.7954 37.1824C17.8215 38.4889 15.3654 40.4486 13.6147 41.4415C13.4579 41.5199 13.275 41.6244 13.1182 41.7028C12.5434 42.0425 11.6811 42.4083 10.662 42.4083C10.4008 42.4083 9.46009 42.4344 8.28426 41.9902C6.32455 41.2325 4.70452 39.3773 3.68547 36.5553C2.69255 33.8639 2.61416 31.1203 3.4503 28.3767C3.68547 27.5406 4.26032 26.3125 4.23419 26.2863C5.51454 23.5427 7.63103 20.982 10.3485 18.8133C10.4008 18.7872 14.4247 15.6516 17.7693 13.4306L13.2227 10.3473H10.0872C7.39586 10.6086 5.0442 11.6277 3.16288 13.4306C1.25543 15.1813 0.105727 18.4475 0.340893 21.583C0.445411 22.9156 0.758965 24.1437 1.28155 25.2934C1.59511 24.6663 2.01318 23.8563 2.5619 22.9156C2.43125 22.393 2.35286 21.8966 2.32674 21.4262C2.14383 18.9178 3.03223 16.2265 4.49548 14.8939C6.06326 13.4045 7.99684 12.5683 10.2178 12.3332L15.8357 12.307C12.6218 14.5019 9.17267 17.1933 9.12041 17.2194C6.87327 19.0223 5.01807 21.0343 3.65934 23.2292C2.79707 24.6663 1.96092 26.5215 1.98705 26.5476C0.628317 30.1013 0.576058 33.7855 1.83027 37.2346C3.39805 41.5199 5.82809 43.1922 7.6049 43.8454C8.46717 44.159 9.40783 44.3158 10.3485 44.3158C11.7856 44.3158 13.0398 43.8454 13.0137 43.8454C13.4318 43.6886 14.2157 43.4012 14.6076 43.166C14.9734 42.957 17.4296 41.546 18.8145 39.6908C20.1993 41.4938 22.5771 42.8525 23.0213 43.1138C23.0213 43.1138 24.0926 43.7409 24.9549 44.0022C25.9217 44.2635 26.9669 44.2635 27.3065 44.2635C28.2472 44.2635 29.1617 44.0806 30.0501 43.767C31.827 43.0877 34.257 41.3892 35.7725 37.104C36.8699 34.0468 36.9222 30.8329 35.9815 27.6712C35.5635 28.2199 35.0409 28.8209 34.3615 29.4219ZM14.477 9.74633L20.8264 14.2929C20.9048 14.3451 21.3229 14.6064 21.3229 14.6064C24.2233 16.5923 27.0453 18.7349 27.0714 18.761C29.4753 20.6162 31.3827 22.8111 32.7154 25.1628C33.2641 24.3789 33.5776 23.5166 33.7344 22.9679C32.3495 20.8514 30.5205 18.8917 28.2995 17.1671C28.2733 17.141 25.7649 15.2335 23.0213 13.3261C21.8716 12.6206 18.4748 10.3212 18.4748 10.3212C16.8547 9.276 14.5292 7.13337 14.5292 7.13337C13.6669 6.29723 13.5624 5.38269 14.6076 5.38269H16.6457V3.39684H14.6337C11.3675 3.42297 10.8711 6.11432 12.6479 8.02178C13.3273 8.77954 14.477 9.74633 14.477 9.74633ZM23.5178 44.5509C22.3158 45.5438 20.4606 45.9619 18.8406 45.988C17.2467 45.9619 15.4437 45.57 14.2418 44.6554C13.6931 44.8645 12.8831 45.1519 12.073 45.3348C12.8047 46.2232 13.8237 46.9026 15.1302 47.3729C16.4628 47.8432 17.7954 47.9739 18.6577 48H19.1019C19.9642 47.9739 21.2968 47.8171 22.6294 47.3468C23.9358 46.8764 24.9549 46.1709 25.6865 45.2564C24.7981 45.0474 24.0142 44.7599 23.5178 44.5509ZM34.2309 13.3C32.3495 11.5754 30.0763 10.5563 27.4633 10.295L27.2282 10.2689H26.8623L22.2635 10.295C22.7339 9.92924 23.2042 9.56342 23.6745 9.17148C23.6745 9.17148 23.7529 9.14535 23.962 8.93631C24.171 8.75341 24.5368 8.3092 24.5368 8.3092C24.6413 8.20468 24.7459 8.07404 24.8765 7.94339C26.6794 6.03593 26.1568 3.34458 22.8907 3.34458H20.8787V5.33043H22.9168C24.0142 5.33043 23.6745 6.2711 22.9168 7.23789C21.8977 8.1263 20.8003 8.96244 19.6245 9.77246C21.0093 10.7654 22.4203 11.6538 23.4394 12.2809L27.1498 12.2548C29.3708 12.4638 31.3044 13.2999 32.8983 14.7632C34.3615 16.0958 35.2761 18.761 35.1193 21.2695C35.067 22.0272 34.9102 22.9418 34.5706 23.8563C34.2047 24.8231 33.5515 26.1296 32.4018 27.0441L32.3757 27.0702C32.0882 27.3315 31.8008 27.5667 31.4611 27.8019C26.2091 31.4861 25.8172 36.1895 25.6081 38.6979C25.582 38.9069 25.582 39.116 25.5559 39.2989C25.5559 39.4034 25.5297 39.534 25.5297 39.6386C25.5036 39.9521 25.4775 40.3441 25.3991 40.736C25.8956 40.9973 26.601 41.2325 27.3588 41.1541C27.4372 40.6315 27.4895 40.1873 27.5156 39.7953C27.5156 39.6908 27.5417 39.5602 27.5417 39.4818C27.5678 39.2727 27.5678 39.0637 27.594 38.8285C27.803 36.3462 28.1427 32.5574 32.5586 29.4219L32.6631 29.3174C35.0931 27.5667 36.0338 25.4502 36.0338 25.4241C36.0599 25.3718 36.0861 25.3457 36.1122 25.3196C36.7132 24.0915 37.0529 22.7327 37.1574 21.3217C37.3142 18.2907 36.1383 15.0245 34.2309 13.3ZM12.2821 40.9189C12.2037 40.5008 12.1776 40.0828 12.1253 39.7692C12.1253 39.6386 12.0992 39.534 12.0992 39.4295C12.073 39.2466 12.0469 39.0376 12.0469 38.8285C11.8117 36.3201 11.3675 31.6168 6.08939 27.9848C5.72357 27.7496 5.41002 27.4883 5.12259 27.227C4.80904 28.0109 4.6 28.7425 4.49548 29.2912C4.65226 29.3958 4.78291 29.5264 4.93969 29.6309C9.46009 32.7403 9.79978 36.5291 10.0349 39.0114C10.0611 39.2466 10.0872 39.4556 10.0872 39.6647C10.0872 39.7692 10.1133 39.8737 10.1133 39.9782C10.1395 40.3441 10.1917 40.8144 10.2962 41.3108C11.0801 41.3892 11.7595 41.1802 12.2821 40.9189Z"
+      fill="#24292F"
+    ></path>
+  </svg>,
+  <svg key={8} role="img" width="71" height="31" viewBox="0 0 71 31" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <title>P&G</title>
+    <path
+      d="M28.1268 30.9843C25.7042 30.8536 23.8966 29.8533 22.8394 28.0586C22.1301 26.8538 21.9143 25.2692 22.2387 23.6502C22.5983 21.8561 23.8192 20.2395 25.6929 19.0768C26.7696 18.4081 27.6912 17.9931 29.5677 17.3312C30.0395 17.1649 30.4467 17.0117 30.473 16.9903C30.5106 16.9604 30.5159 16.8542 30.4971 16.5098C30.4431 15.5298 30.6061 14.0029 30.8819 12.9011C31.3459 11.0469 32.2387 9.55105 33.4237 8.64372C34.2743 7.9918 35.3894 7.50162 36.4994 7.29145C37.1125 7.17567 38.1392 7.17602 38.7161 7.29251C41.1879 7.79196 42.5753 9.70387 42.0289 11.858C41.9139 12.3122 41.5504 13.0568 41.2426 13.4686C40.4541 14.5237 39.2681 15.4621 37.5932 16.3548C37.3827 16.4674 36.6933 16.8072 36.0617 17.1107C35.4302 17.4139 34.8938 17.6821 34.8689 17.7071C34.8338 17.7423 34.8427 17.8453 34.9108 18.1915C35.225 19.7925 35.7532 21.1451 36.6553 22.6577C37.0138 23.2597 37.5098 24.0189 37.5662 24.0541C37.6479 24.1047 37.9816 23.5636 38.1992 23.0282C38.4683 22.3656 38.6223 21.4892 38.5464 21.0539C38.4917 20.7422 38.322 20.413 38.1055 20.2004C38.0061 20.102 37.9237 20.0091 37.9233 19.9937C37.923 19.9788 38.7636 19.9702 39.7921 19.9756L41.6618 19.9852L41.6512 20.2232C41.6366 20.5484 41.5163 21.242 41.3896 21.7357C41.0321 23.124 40.3583 24.3113 39.1935 25.608C39.0561 25.7608 38.9436 25.9108 38.9436 25.9415C38.9436 25.9721 39.1367 26.1848 39.3724 26.4142C40.8446 27.8463 43.4525 29.4978 45.5757 30.3424C45.782 30.4247 45.9502 30.5084 45.9495 30.5284C45.9488 30.5526 44.723 30.5683 42.3672 30.5747L38.786 30.5839L38.5822 30.4689C38.1662 30.2341 37.2449 29.4871 36.3599 28.6663C36.1725 28.4925 35.995 28.3503 35.9655 28.3503C35.9361 28.3503 35.7277 28.4718 35.5033 28.6204C34.3535 29.3813 32.9793 30.0649 31.817 30.4539C30.495 30.8967 29.4037 31.0534 28.1268 30.9847V30.9843ZM31.1421 27.8708C31.6782 27.775 32.3367 27.5798 32.9423 27.3372C33.5512 27.0935 34.3059 26.7209 34.3329 26.6504C34.3435 26.6226 34.204 26.3885 34.0233 26.1299C33.1901 24.9387 32.3001 23.2633 31.702 21.7607C31.4815 21.2067 31.0697 19.9799 30.9909 19.6436C30.9661 19.5363 30.921 19.4391 30.8908 19.4273C30.8074 19.3953 30.1233 19.7611 29.5322 20.1544C27.4313 21.5519 26.4242 23.2579 26.6478 25.0405C26.7472 25.8328 26.9972 26.4103 27.4519 26.898C28.2925 27.7992 29.6068 28.1458 31.1421 27.8708ZM34.7081 15.4753C35.4856 15.0186 36.7188 13.9096 37.4086 13.0475C38.2915 11.9442 38.7608 10.9147 38.7611 10.0797C38.7611 9.60235 38.6937 9.38861 38.4736 9.16774C38.2546 8.9483 38.0075 8.85746 37.6301 8.8571C36.5104 8.8571 35.2995 10.1908 34.772 12.0055C34.5058 12.9203 34.4411 13.4126 34.4394 14.5302C34.438 15.3837 34.4518 15.5736 34.5157 15.5736C34.5295 15.5736 34.6162 15.5294 34.7081 15.4753ZM53.226 30.963C51.381 30.7766 50.1652 30.4999 48.7934 29.9548C46.9268 29.2124 45.7614 28.293 44.777 26.7864C44.4823 26.3355 44.0095 25.3886 43.8185 24.8653C42.9647 22.5337 42.7773 19.5983 43.286 16.5361C43.8011 13.4369 44.9743 10.4498 46.5931 8.11506C47.6432 6.60106 49.2056 5.00655 50.8184 3.80247C53.4816 1.81468 56.547 0.592787 59.9525 0.163167C60.9856 0.0327848 61.5642 -0.000701409 62.7843 1.10625e-05C64.63 0.00107977 66.0358 0.142149 67.8921 0.512278C69.0764 0.748819 70.9628 1.25966 70.999 1.35406C71.0175 1.4018 69.8222 5.24594 69.7729 5.29759C69.7576 5.31362 69.5741 5.19072 69.3657 5.02329C67.6869 3.68028 65.947 2.86272 64.1877 2.59091C63.6257 2.50399 62.5696 2.50292 62.0186 2.58878C59.5521 2.97209 57.2652 4.55342 55.2349 7.27969C54.5753 8.16529 53.8486 9.35227 53.2835 10.4666C52.5177 11.977 51.9913 13.3122 51.4588 15.0976C49.7626 20.7782 49.6767 24.747 51.2032 26.8955C51.8319 27.7807 52.751 28.3446 53.9172 28.5601C54.5867 28.6841 56.2119 28.6089 56.3738 28.4469C56.3997 28.4212 57.1931 25.9301 58.1367 22.911C59.6873 17.9518 59.8549 17.3925 59.8726 17.1114C59.8961 16.7338 59.8261 16.3965 59.6568 16.0716C59.5031 15.7763 59.097 15.3755 58.7622 15.1892C58.6177 15.1087 58.4991 15.0303 58.4988 15.0154C58.4984 15.0004 61.0999 14.9879 64.2803 14.9879C67.4604 14.9879 70.0625 15.0004 70.0622 15.0154C70.0622 15.0303 69.9213 15.1076 69.7491 15.1863C68.8488 15.5996 68.1963 16.2187 67.8083 17.0288C67.7242 17.2041 66.9964 19.4726 65.9034 22.9658L64.1344 28.6193L63.6232 28.8651C61.5411 29.8679 58.994 30.59 56.4533 30.8981C55.9864 30.9551 53.6168 31.0025 53.2263 30.963H53.226ZM0.00117775 30.5497C-0.0105373 30.5312 0.0664977 30.466 0.171933 30.4051C0.800283 30.0421 1.50567 29.289 1.80458 28.6613C1.8912 28.4796 9.82296 3.1299 9.99798 2.47478C10.0289 2.36007 10.0548 2.11783 10.0551 1.93686C10.0576 1.38399 9.88651 0.985359 9.48358 0.602762C9.37353 0.498385 9.28372 0.39757 9.28372 0.37869C9.28372 0.32846 23.134 0.332735 23.7691 0.383321C25.1444 0.492685 26.5211 0.859608 27.574 1.39681C29.3717 2.31412 30.4481 3.70629 30.779 5.54197C30.8713 6.05281 30.889 7.18921 30.8134 7.73923C30.5053 9.98138 29.4928 12.3596 28.1143 14.0802C25.8814 16.8674 22.2295 18.6497 17.7476 19.1398C16.3269 19.2952 14.5367 19.3344 13.2689 19.2375C12.9026 19.2093 12.5952 19.194 12.5859 19.2033C12.5408 19.2489 9.68132 28.4558 9.64866 28.6613C9.54784 29.2951 9.87302 30.0075 10.4336 30.3809C10.5443 30.4546 10.6253 30.5305 10.6139 30.549C10.5855 30.595 0.0292227 30.5953 0.000822747 30.5494L0.00117775 30.5497ZM15.7227 16.8749C16.7639 16.6897 17.6827 16.3028 18.5691 15.6769C18.9997 15.373 19.8844 14.4863 20.2493 13.993C21.1624 12.7593 21.834 11.3037 22.3758 9.38647C22.7208 8.16458 22.8678 7.23374 22.8674 6.27439C22.8674 5.53485 22.8007 5.09311 22.6154 4.59759C22.3403 3.86232 21.7801 3.2813 21.0683 2.99275C20.457 2.74481 19.9351 2.67962 18.6799 2.69422L17.7334 2.70491L15.5335 9.73415C14.324 13.6004 13.3407 16.7823 13.3488 16.8051C13.3573 16.8279 13.5093 16.8664 13.6861 16.8906C13.8636 16.9148 14.0329 16.9404 14.0631 16.9479C14.2008 16.9818 15.427 16.928 15.7227 16.8753V16.8749Z"
+      fill="#24292F"
+    ></path>
+  </svg>,
+  <svg key={9} role="img" width="48" height="48" viewBox="0 0 48 48" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <title>Dell</title>
+    <path
+      d="M24.0016 2.73684C18.1319 2.73684 12.8162 5.11769 8.96761 8.96925C5.11751 12.812 2.73813 18.1306 2.73813 24.0033C2.73813 29.8715 5.11751 35.1864 8.96761 39.0307C12.8162 42.8771 18.1319 45.2624 24.0016 45.2624C29.8728 45.2624 35.1914 42.8771 39.0342 39.0307C42.8828 35.1864 45.2571 29.8715 45.2571 24.0033C45.2571 18.1306 42.8828 12.812 39.0342 8.96925C35.1914 5.11769 29.8728 2.73684 24.0016 2.73684ZM24.0016 0.160332C37.1663 0.160332 47.8402 10.8357 47.8402 24.0033C47.8402 37.1664 37.1663 47.8403 24.0016 47.8403C10.8384 47.8403 0.160156 37.1664 0.160156 24.0033C0.160156 10.8357 10.8384 0.160332 24.0016 0.160332Z"
+      fill="#24292F"
+    ></path>
+    <path
+      d="M12.7255 23.8086C12.7255 22.4192 11.8036 21.63 10.4985 21.63H9.69975V25.9887H10.4758C11.6893 25.9887 12.7255 25.3292 12.7255 23.8086ZM26.9336 25.3768L21.0493 30.0103L15.6713 25.7762C14.8953 27.6404 13.02 28.9352 10.8605 28.9352H6.25781V18.682H10.8605C13.2736 18.682 15.0125 20.2164 15.6698 21.8513L21.0493 17.6216L23.0417 19.1898L18.1115 23.0648L19.059 23.8064L23.9878 19.9336L25.9832 21.4974L21.0537 25.3768L21.9997 26.1198L26.9336 22.2397V18.682H30.5104V25.9425H34.0784V28.9477H26.9336V25.3768ZM38.6092 25.944H42.1757V28.9477H35.0288V18.682H38.6092V25.944Z"
+      fill="#24292F"
+    ></path>
+  </svg>,
+  <svg key={10} role="img" width="107" height="49" viewBox="0 0 107 49" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <title>Ford</title>
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M0 24.4976C0 13.373 23.7329 4.5 53.3333 4.5C82.743 4.5 106.667 13.4684 106.667 24.4976C106.667 35.5268 82.743 44.5 53.3333 44.5C23.9284 44.5 0 35.5555 0 24.4976ZM74.5899 28.9578C74.4754 28.9148 74.3561 28.829 74.3418 28.7336C74.3244 28.673 74.3313 28.608 74.3609 28.5523C74.7902 27.6793 75.0955 26.5678 74.1319 26.1051C73.2064 25.6614 71.7896 26.4819 70.5064 28.0562C69.4569 29.349 68.8844 30.6608 68.8654 31.9441C68.8606 32.3973 69.0323 32.9602 69.433 33.1701C70.3394 33.6423 71.4318 32.9363 72.2237 32.2828C72.7437 31.8534 73.1778 31.4098 73.5595 30.8373C73.6179 30.75 73.6741 30.6667 73.7285 30.5863C74.0598 30.0957 74.3221 29.7074 74.6089 29.1009C74.6567 29.0007 74.6376 28.9721 74.5899 28.9578ZM84.5076 18.5487C84.665 18.5487 84.7318 18.6393 84.665 18.7634V18.7586C84.5522 18.9515 83.8796 19.991 82.9535 21.4223L82.9534 21.4225L82.9528 21.4234C80.5117 25.1963 76.3104 31.6898 75.9542 32.5785C75.7968 32.9745 75.7777 33.2703 75.9399 33.4849C76.3502 34.0335 77.2565 33.7139 77.829 33.318C79.0432 32.4776 79.991 31.514 80.9119 30.5778L80.912 30.5777L80.9121 30.5776L80.9122 30.5775C81.272 30.2117 81.6277 29.8501 81.9936 29.5016C82.0651 29.4491 82.1939 29.4253 82.2798 29.5016C82.3534 29.5716 82.4406 29.6487 82.5309 29.7285L82.531 29.7286L82.5311 29.7287L82.5312 29.7288L82.5313 29.7289C82.6747 29.8557 82.8259 29.9893 82.9429 30.1122C83.0383 30.2124 83.0001 30.3078 82.9524 30.3651C81.6835 31.9154 80.281 33.38 78.9357 34.4152C76.8463 36.0323 74.7616 36.4474 73.7312 35.4122C73.3496 35.021 73.1397 34.4724 73.111 34.0049C73.0967 33.7521 72.9679 33.7378 72.8487 33.857C71.6131 35.0782 69.4712 36.4235 67.5249 35.4694C65.8934 34.6584 65.5165 32.7169 65.8218 31.0806C65.8314 31.0472 65.7932 30.8135 65.5928 30.9518C65.2589 31.1951 64.8105 31.4623 64.2571 31.6674C63.7488 31.9012 63.1947 32.0185 62.6352 32.0109C62.2319 32.015 61.8346 31.9131 61.483 31.7156C61.1314 31.518 60.8378 31.2316 60.6316 30.885C60.3454 30.4605 60.2357 29.8165 60.2023 29.349C60.2023 29.2965 60.1355 29.2011 60.0592 29.3251C58.991 30.9522 57.9813 32.5394 56.9436 34.1704C56.6882 34.5717 56.4312 34.9757 56.1713 35.3836C56.1045 35.479 56.0425 35.541 55.9041 35.5505H53.2995C53.1373 35.5505 53.0609 35.4265 53.1468 35.2738C53.323 34.9358 53.7919 34.2021 54.3998 33.2509L54.4 33.2506L54.4 33.2506L54.4001 33.2506C55.4051 31.6779 56.79 29.511 57.86 27.5553C58.3323 26.6918 58.2369 26.2482 57.9411 26.0574C57.5738 25.8188 56.8439 26.1766 56.2237 26.6489C54.7999 27.7258 53.5671 29.3143 52.7877 30.3186L52.7876 30.3186L52.787 30.3195C52.6448 30.5026 52.5178 30.6663 52.4074 30.8039C52.2686 30.9739 52.1323 31.1714 51.9733 31.402L51.9733 31.402L51.9733 31.4021C51.6458 31.8767 51.2218 32.4914 50.4801 33.2941C48.7437 35.1832 46.4396 36.2661 44.2213 35.5171C43.6369 35.3183 43.1103 34.979 42.6877 34.529C42.2652 34.079 41.9596 33.5321 41.798 32.9363C41.7312 32.7073 41.5976 32.6406 41.4736 32.7932C41.1158 33.2416 40.3812 33.8999 39.7705 34.315C39.7038 34.3627 39.5368 34.4629 39.3937 34.3913C39.1272 34.1301 38.9006 33.8311 38.721 33.504C38.694 33.4452 38.6864 33.3793 38.6992 33.3158C38.7121 33.2524 38.7448 33.1946 38.7926 33.151C39.6704 32.3829 40.1474 31.9441 40.529 31.467C40.8518 31.0409 41.1076 30.5679 41.2875 30.0645C41.8171 28.7193 43.0335 26.21 45.4521 24.9124C46.3776 24.4211 47.3126 24.0633 48.3525 23.9441C48.7962 23.8963 49.1635 23.9488 49.2971 24.3018L49.5499 24.9983C49.5976 25.1414 49.5595 25.2512 49.3925 25.3132C47.9661 25.857 46.535 26.6966 45.5428 28.0228C44.4217 29.5302 44.1212 31.3382 44.8701 32.3782C45.4187 33.1367 46.306 33.2702 47.1552 32.9983C48.4479 32.5928 49.5451 31.3859 50.1128 30.0502C50.7282 28.5905 50.8284 26.8636 50.3752 25.3466C50.2989 25.0842 50.3657 24.8838 50.5374 24.7407C50.8314 24.4998 51.2197 24.3079 51.4122 24.2127C51.4447 24.1966 51.4715 24.1833 51.4915 24.173C51.627 24.1 51.7847 24.0796 51.9344 24.1157C52.084 24.1518 52.2151 24.2419 52.3024 24.3686C52.7969 25.1838 53.0561 26.12 53.0514 27.0735C53.0848 27.2881 53.1754 27.2929 53.2756 27.1784C53.4936 26.9469 53.7077 26.7113 53.9209 26.4767C54.5337 25.8024 55.1386 25.1368 55.8039 24.6024C56.6245 23.9488 57.6787 23.4145 58.6614 23.4384C59.9304 23.467 60.5648 24.3305 60.5219 25.8665C60.5171 26.0574 60.6555 26.0478 60.7509 25.9142C61.0449 25.4707 61.4053 25.0749 61.8194 24.7407C62.6972 24.0156 63.5988 23.6912 64.5338 23.758C64.8152 23.7775 65.0895 23.8547 65.3397 23.985C65.5899 24.1152 65.8105 24.2957 65.9878 24.5151C66.1651 24.7344 66.2952 24.988 66.3701 25.2599C66.4449 25.5319 66.4629 25.8163 66.4229 26.0955C66.3764 26.4749 66.2384 26.8373 66.0206 27.1515C65.8029 27.4656 65.512 27.7221 65.173 27.8987C64.5672 28.1993 63.866 28.1468 63.3746 27.7079C63.1832 27.5284 63.0427 27.3015 62.9673 27.0501C62.8919 26.7988 62.8843 26.532 62.9453 26.2768C62.9691 26.1671 62.826 26.0621 62.7067 26.1862C61.5762 27.3597 61.4951 28.8815 62.2297 29.5636C63.0454 30.3174 64.2619 29.8117 65.154 29.2345C66.1026 28.6132 66.9254 27.8661 67.7349 27.131C68.3177 26.6019 68.8936 26.0789 69.5046 25.6137C70.4778 24.8791 71.5797 24.2255 72.82 24.0538C73.5019 23.9518 74.1986 24.0342 74.8379 24.2923C75.7252 24.6501 76.2881 25.399 76.3597 26.0478C76.374 26.2386 76.5219 26.2625 76.6364 26.0955C76.881 25.7128 77.4987 24.7771 78.2127 23.6956L78.2129 23.6952C79.546 21.6759 81.2144 19.1486 81.4163 18.7634C81.507 18.5773 81.6119 18.5487 81.8123 18.5487H84.5076ZM47.9137 20.1992C47.9471 20.3757 48.0186 20.4091 48.0997 20.3901C48.5447 20.2689 48.931 19.9911 49.1874 19.6077C49.4259 19.2642 49.5356 18.7681 49.3019 18.5678C49.092 18.3912 48.7533 18.439 48.5004 18.6632C48.2856 18.8526 48.12 19.0914 48.0178 19.3589C47.9156 19.6265 47.8799 19.9148 47.9137 20.1992ZM28.0067 35.5075C25.0967 35.5266 22.5302 33.7282 22.2345 30.6083C22.0642 28.9859 22.5312 27.3609 23.5368 26.0764C24.405 24.9458 25.8743 24.0108 27.1146 23.7818C27.5153 23.7055 28.0592 23.7532 28.2834 24.0203C28.6364 24.4306 28.5362 24.879 28.1832 25.0603C27.2339 25.5421 26.0938 26.3722 25.5308 27.5028C25.2709 27.9832 25.1472 28.5257 25.1733 29.0714C25.1993 29.6171 25.3741 30.1452 25.6787 30.5988C26.8475 32.3638 30.1104 32.4115 33.1397 29.597C33.8934 28.9053 34.5613 28.2422 35.2434 27.4598C36.5791 25.9333 37.0228 25.3751 38.7974 22.9136C38.8165 22.885 38.826 22.8468 38.8165 22.8325C38.8022 22.7991 38.7783 22.7943 38.7258 22.7848C37.5428 22.823 36.5076 23.176 35.5058 23.9249C35.2196 24.1348 34.7998 24.1062 34.6185 23.8295C34.4181 23.5147 34.5613 23.1044 34.8952 22.8754C36.3559 21.8555 37.942 21.5951 39.5682 21.3282C39.737 21.3005 39.9063 21.2727 40.0759 21.244C40.0759 21.244 40.1999 21.244 40.3048 21.1152C40.6119 20.719 40.8877 20.3401 41.1616 19.9637C41.5473 19.4338 41.9292 18.909 42.3895 18.3483C43.1862 17.3751 44.2118 16.4783 44.9942 15.8057C45.9299 15.0042 46.5777 14.6572 47.189 14.3298C47.2945 14.2733 47.3988 14.2174 47.5034 14.1599C47.6942 14.0645 47.4796 13.9452 47.4223 13.9452C47.1298 13.8902 46.8365 13.834 46.5424 13.7777L46.5419 13.7776C42.8322 13.0671 39.0125 12.3355 35.3484 13.6924C32.6101 14.7133 31.408 16.8886 31.9852 18.3388C32.3955 19.3692 33.7503 19.5314 35.0192 18.9112C36.0687 18.4008 37.0466 17.566 37.7384 16.4974C38.12 15.9011 38.9357 16.3304 38.4968 17.3656C37.3376 20.0847 35.1289 22.241 32.4384 22.3316C30.1438 22.4079 28.665 20.7192 28.6698 18.6393C28.6793 14.5367 33.2494 11.8367 38.0437 11.4312C42.0887 11.0975 45.9842 11.7514 49.9247 12.4129C51.6166 12.6969 53.3168 12.9823 55.0407 13.1915C59.439 13.721 63.6561 13.7353 68.0782 12.7144C68.603 12.6047 68.9655 12.9387 68.9369 13.4634C68.8892 14.2267 68.2977 14.9613 66.8427 15.7532C65.1874 16.6644 63.4557 16.941 61.6429 16.9554C58.641 16.9816 55.7787 16.1838 52.9122 15.3848C51.606 15.0208 50.299 14.6565 48.9775 14.3698C48.9784 14.4044 48.9809 14.4402 48.9835 14.4762L48.9835 14.4764C48.9947 14.6341 49.0063 14.7975 48.8821 14.8945C46.5446 16.6119 44.4313 18.3626 43.0144 20.8194C42.9667 20.8766 42.9954 20.9339 43.0765 20.9243C43.4159 20.9072 43.7455 20.8923 44.069 20.8777H44.069C44.9187 20.8392 45.7263 20.8026 46.5589 20.7335C46.7545 20.7192 46.764 20.6715 46.7497 20.5236C46.6877 20.07 46.736 19.6081 46.8904 19.1771C47.0449 18.7462 47.301 18.3588 47.637 18.0478C48.4002 17.3656 49.3543 17.2845 49.9888 17.7806C50.7044 18.3388 50.6424 19.3501 50.2512 20.1038C50.0749 20.4357 49.8337 20.7287 49.542 20.9655C49.2502 21.2024 48.9139 21.3781 48.5529 21.4825C48.5529 21.4825 48.4193 21.5111 48.4384 21.6304C48.4527 21.7443 48.7679 21.962 49.0349 22.1463C49.215 22.2707 49.3732 22.3799 49.402 22.4318C49.4736 22.5558 49.4354 22.7419 49.3066 22.885C49.2556 22.9471 49.1916 22.9972 49.119 23.0319C49.0465 23.0665 48.9672 23.0847 48.8868 23.0853C48.7435 23.0853 48.6539 23.036 48.4748 22.9375L48.4575 22.9279C48.0027 22.6849 47.598 22.3579 47.2649 21.9643C47.1504 21.826 47.0932 21.8212 46.8212 21.8403C45.5809 21.9118 43.6441 22.0979 42.337 22.2601C42.0174 22.2982 41.9697 22.3602 41.9077 22.47L41.7687 22.6992C39.7193 26.0782 37.5314 29.6855 35.1337 32.1205C32.51 34.7872 30.3251 35.4884 28.0067 35.5075ZM4.02592 24.5068C4.02592 15.2951 26.0987 7.77215 53.333 7.77215C80.5626 7.77215 102.64 15.2999 102.64 24.5068C102.64 33.7137 80.5674 41.2414 53.333 41.2414C26.0987 41.2414 4.02592 33.7137 4.02592 24.5068ZM2.22747 24.5068C2.22747 34.5867 25.1064 42.7584 53.333 42.7584C81.5548 42.7584 104.434 34.5867 104.434 24.5068C104.434 14.4221 81.5548 6.25516 53.333 6.25516C25.1064 6.25516 2.22747 14.4221 2.22747 24.5068Z"
+      fill="#24292F"
+    ></path>
+  </svg>,
+  <svg key={11} role="img" xmlns="http://www.w3.org/2000/svg" fill="none" height="40" viewBox="0 0 68 40" width="68">
+    <title>Stripe</title>
+    <path
+      clipRule="evenodd"
+      d="m36.708 5.854v3.80384l-4.7032 1.01186v-3.82258zm9.7812 5.921c-1.8364 0-3.0169.862-3.6727 1.4616l-.2436-1.1618h-4.1224v21.8487l4.6845-.9932.0188-5.3029c.6746.4872 1.6677 1.1805 3.3166 1.1805 3.3541 0 6.4085-2.6982 6.4085-8.6382-.0188-5.4341-3.1106-8.3947-6.3897-8.3947zm-1.1243 12.9106c-1.1056 0-1.7614-.3935-2.2111-.8807l-.0188-6.9519c.4872-.5434 1.1618-.9181 2.2299-.9181 1.7051 0 2.8856 1.9113 2.8856 4.3659 0 2.511-1.1617 4.3848-2.8856 4.3848zm22.2797-4.3285c0-4.797-2.3235-8.5821-6.7644-8.5821-4.4597 0-7.158 3.7851-7.158 8.5446 0 5.6402 3.1855 8.4883 7.7576 8.4883 2.2298 0 3.9162-.5059 5.1904-1.2179v-3.7476c-1.2742.637-2.7357 1.0305-4.5908 1.0305-1.8176 0-3.4291-.6371-3.6352-2.8481h9.1629c0-.1033.0068-.3376.0145-.6077v-.0006-.0001-.0003l.0001-.0007v-.0006-.0004c.0105-.3668.0229-.7988.0229-1.0573zm-9.2566-1.7802c0-2.1174 1.2929-2.9981 2.4734-2.9981 1.1431 0 2.361.8807 2.361 2.9981zm-21.68-6.4829h-4.7032v16.3959h4.7032zm-10.0434.0001.2998 1.3867c1.1056-2.0238 3.2979-1.6115 3.8976-1.3867v4.3098c-.5809-.2061-2.4547-.4685-3.5603.9744v11.1117h-4.6845v-16.3959zm-9.0694-4.06648-4.5721.97438-.0188 15.0092c0 2.7733 2.08 4.8157 4.8532 4.8157 1.5365 0 2.6608-.2811 3.2792-.6183v-3.8039c-.5996.2436-3.5603 1.1056-3.5603-1.6677v-6.652h3.5603v-3.9912h-3.5603zm-11.07422 7.81358c-.99312 0-1.59274.2811-1.59274 1.0118 0 .7979 1.03195 1.1489 2.31221 1.5844 2.08711.7098 4.83415 1.6442 4.84575 5.1051 0 3.3542-2.67956 5.2842-6.57708 5.2842-1.61148 0-3.37286-.3186-5.115501-1.0681v-4.4597c1.574001.862 3.560241 1.4991 5.115501 1.4991 1.04933 0 1.79886-.2811 1.79886-1.143 0-.8839-1.11866-1.2878-2.46918-1.7756-2.05676-.7427-4.6513-1.6797-4.6513-4.8015 0-3.3166 2.52964-5.3029 6.33348-5.3029 1.55526 0 3.09178.2436 4.64702.862v4.4034c-1.42407-.7682-3.22293-1.1992-4.64702-1.1992z"
+      fill="#474f5d"
+      fillRule="evenodd"
+    ></path>
+  </svg>,
+  <svg key={12} role="img" width="60" height="32" viewBox="0 0 60 32" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <title>3m</title>
+    <path
+      d="M42.2721 19.5456L36.7445 0H24.3342V5.23796C22.7171 2.09889 18.8113 0.215449 13.3648 0.215449C6.96619 0.215449 1.26026 3.22015 0.954463 10.2744H9.14153C9.40331 8.09672 10.9694 7.01021 13.103 7.01021C14.8868 7.01021 16.6707 7.7469 16.6707 9.75313C16.6707 12.6235 13.6683 12.8853 11.2729 12.8853V18.2437C13.4482 18.2437 16.9325 18.5472 16.9325 21.6399C16.9325 23.4237 15.5378 25.2099 12.8853 25.2099C10.0566 25.2099 8.53225 23.5094 8.53225 20.6831H0C0 28.2146 5.26808 32 12.408 32C17.968 32 22.1658 30.3088 24.3365 27.0377V31.3513H33.5684V20.7248C33.5684 18.1973 33.2186 13.9764 32.9151 11.2335H33.0031L38.7044 31.349H45.8443L51.5456 11.2335H51.6337C51.2422 14.8451 50.9804 18.2437 50.9804 21.7673V31.3467H60.2099V0H47.8042L42.2721 19.5456ZM24.3342 16.7101C23.623 15.9154 22.7056 15.2784 21.5936 14.8451C22.8585 14.2243 23.755 13.381 24.3342 12.3408V16.7101Z"
+      fill="#24292F"
+    ></path>
+  </svg>,
+  <svg key={13} role="img" width="106" height="23" viewBox="0 0 106 23" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <title>Mercari</title>
+    <g>
+      <path
+        d="M30.3806 8.76807H19.9531V22.7806H30.3806V19.9142H23.0728V17.2524H30.0429V14.426H23.0728V11.6344H30.3806V8.76807Z"
+        fill="#24292F"
+      ></path>
+      <path
+        d="M46.9788 10.9422C46.9788 9.51488 46.4622 8.37697 45.4407 7.56015C44.4459 6.74831 43.1452 6.3374 41.577 6.3374H35.7422V20.3482H38.8619V15.9031H41.1005L43.8407 20.3482H47.3851L44.2771 15.3075C45.0986 14.9117 45.7894 14.2905 46.268 13.5171C46.7467 12.7438 46.9933 11.8504 46.9788 10.9422ZM38.8619 9.20212H41.0654C42.9897 9.20212 43.8859 9.806 43.8859 11.1003C43.8859 12.3946 42.9663 13.0567 41.0654 13.0567H38.8619V9.20212Z"
+        fill="#24292F"
+      ></path>
+      <path
+        d="M62.2092 18.1872C61.2713 19.4499 60.0241 20.0921 58.4993 20.0921C57.9492 20.1058 57.4019 20.0099 56.8897 19.8099C56.3774 19.6099 55.9106 19.3099 55.5167 18.9275C54.7159 18.1506 54.3096 17.0793 54.3096 15.7434C54.3096 14.4075 54.7159 13.3778 55.5167 12.6192C55.8857 12.2476 56.3255 11.9532 56.8105 11.7532C57.2954 11.5533 57.8156 11.4517 58.3405 11.4546C59.0508 11.4434 59.7531 11.6055 60.3858 11.9268C61.0186 12.2482 61.5626 12.7189 61.9701 13.2979L62.0253 13.3761L64.1285 11.2417L64.0834 11.1868C62.664 9.4367 60.7798 8.54834 58.4809 8.54834C56.3827 8.54834 54.6139 9.21378 53.2212 10.5447C51.8286 11.8755 51.1113 13.629 51.1113 15.8066C51.1113 17.9543 51.8219 19.7094 53.2229 21.0237C54.6239 22.3379 56.4429 23 58.5729 23C60.9302 23 62.8562 22.1133 64.2957 20.3632L64.3392 20.3083L62.2594 18.114L62.2092 18.1872Z"
+        fill="#24292F"
+      ></path>
+      <path
+        d="M75.6076 6.3374H72.6785L67.2098 20.2401L67.168 20.3482H70.398L71.4012 17.8379H76.8498L77.8746 20.3482H81.103L75.6193 6.38731L75.6076 6.3374ZM75.8383 15.0696H72.4494L74.143 10.6311L75.8383 15.0696Z"
+        fill="#24292F"
+      ></path>
+      <path
+        d="M96.4077 13.3729C96.4077 11.9455 95.8894 10.8076 94.8696 9.99081C93.8497 9.17399 92.5741 8.76807 91.0059 8.76807H85.166V22.7806H88.2857V18.3338H90.5244L93.2646 22.7806H96.8089L93.7009 17.7332C94.5226 17.3384 95.2139 16.7183 95.6934 15.9459C96.1729 15.1734 96.4207 14.2808 96.4077 13.3729ZM88.2857 11.6344H90.4893C92.4136 11.6344 93.308 12.2367 93.308 13.531C93.308 14.8252 92.3868 15.4873 90.4893 15.4873H88.2857V11.6344Z"
+        fill="#24292F"
+      ></path>
+      <path d="M104.574 7.55176H101.475V20.3482H104.574V7.55176Z" fill="#24292F"></path>
+      <path
+        d="M102.603 0.000262425L101.17 3.37988L104.566 4.8061L106 1.42648L102.603 0.000262425Z"
+        fill="#24292F"
+      ></path>
+      <path
+        d="M7.25424 13.9217L11.2433 6.3374H14.4199V20.3482H11.2851V12.348H11.2483L8.22727 17.7563H6.19929L3.17655 12.3563H3.1381V20.3416H0V6.3374H3.17655L7.18904 13.9217H7.25424Z"
+        fill="#24292F"
+      ></path>
+    </g>
+    <defs>
+      <clipPath>
+        <rect width="106" height="23" fill="white"></rect>
+      </clipPath>
+    </defs>
+  </svg>,
+  <svg key={14} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 41 41" width="41" height="41" fill="none">
+    <title>Microsoft</title>
+    <path
+      fill="#8B949E"
+      d="m.816 6.404 16.346-2.22.008 15.72-16.34.093L.817 6.404Zm16.339 15.311.012 15.733-16.339-2.24V21.61l16.327.106Zm1.981-17.82L40.811.74v18.963l-21.675.171V3.895Zm21.68 17.968L40.81 40.74l-21.675-3.05-.03-15.863 21.71.035Z"
+    />
+  </svg>,
+  <svg key={15} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 151 28" width="151" height="28" fill="none">
+    <title>BuzzFeed</title>
+    <path
+      fill="#8B949E"
+      d="M.155.474C4.378.48 8.602.417 12.823.514c3.254.074 5.803 2.284 6.352 5.302.62 3.417-1.152 6.335-3.887 7.118-.07.02-.126.084-.22.15 2.607.746 4.164 2.471 4.602 5.057.722 4.265-1.63 7.77-6.076 8.266-.986.11-1.986.107-2.98.113-3.487.018-6.973.017-10.459.024V.474ZM5.77 21.567c.073.025.145.071.217.071 1.75.003 3.503.065 5.249-.017 1.806-.086 2.88-1.197 2.929-2.833.048-1.644-.945-2.83-2.714-3.042-1.118-.134-2.26-.073-3.393-.089-.75-.01-1.5-.002-2.288-.002v5.912Zm-.038-10.762c1.631 0 3.18.002 4.728-.003.298 0 .603.002.893-.057 1.094-.22 1.917-.78 2.204-1.92.456-1.817-.702-3.387-2.582-3.48-.874-.044-1.752-.012-2.629-.013H5.77c-.017.188-.037.303-.037.417-.002 1.651-.001 3.302-.001 5.056Zm95.115 16.217c-.66-.122-1.327-.216-1.98-.373-5.105-1.224-7.707-5.618-7.26-10.474.373-4.066 2.423-7.038 6.326-8.414 5.451-1.921 11.771.673 12.704 7.925.123.958.117 1.932.176 2.986H96.736c.327 1.786 1.228 3.04 2.799 3.706 2.553 1.082 4.944.543 7.199-.909.133-.085.256-.187.386-.28.012-.009.038 0 .128 0l2.202 3.242c-.726.453-1.395.944-2.127 1.311-1.346.676-2.793 1.035-4.296 1.16-.172.014-.338.079-.507.12h-1.674Zm5.107-11.782c.098-2.341-2.132-4.13-4.957-4.018-2.35.093-4.438 2.082-4.187 4.018h9.144Zm15.103 11.782c-.813-.163-1.638-.279-2.435-.498-3.805-1.05-6.43-4.268-6.775-8.2-.326-3.71.749-6.854 3.727-9.195 4.513-3.546 13.381-2.741 15.143 5.65.126.6.2 1.218.238 1.831.041.651.009 1.308.009 2.061H116.98c.122 1.049.529 1.892 1.205 2.59 1.547 1.596 3.477 1.896 5.562 1.53 1.329-.233 2.556-.738 3.66-1.683l2.251 3.3c-.586.391-1.117.806-1.7 1.125-1.45.793-3.019 1.212-4.662 1.366-.191.017-.377.08-.566.123h-1.674Zm-4.042-11.785h9.153c.118-2.27-2.06-4.073-4.797-4.02-2.429.048-4.531 1.969-4.356 4.02ZM28.497 27.021c-.598-.122-1.2-.228-1.794-.37-3.677-.876-5.85-3.629-5.852-7.412-.003-3.626 0-7.253 0-10.88v-.65h4.97v.693c0 3.427.01 6.855-.005 10.282-.007 1.555.469 2.871 1.942 3.548 1.596.733 3.231.706 4.708-.346 1.082-.77 1.388-1.95 1.39-3.216.007-3.408.002-6.815.002-10.223V7.7h4.917v.672c0 3.587.002 7.173 0 10.76-.003 4.328-2.697 7.304-7.014 7.757-.214.023-.422.088-.633.134h-2.631Zm46.93-.538V.492h18.078v2.521c0 .558-.037 1.12.01 1.673.048.552-.173.659-.68.656-3.568-.02-7.135-.01-10.702-.01h-1.129v5.473h.674c2.69 0 5.38.015 8.07-.013.545-.006.672.167.66.68-.034 1.37-.012 2.741-.012 4.18h-9.432v10.832h-5.537Zm56.796-.01V8.34c0-.764.001-.77.789-.763 2.728.025 5.467-.091 8.18.12 3.867.298 6.893 2.04 8.326 5.857 1.456 3.878.346 8.305-2.612 10.73-1.743 1.429-3.794 2.132-6.003 2.212-2.805.101-5.617.032-8.425.034-.057 0-.114-.024-.255-.056Zm4.74-3.99h2.736c3.21 0 5.231-1.809 5.591-5.003.258-2.292-.983-4.573-2.958-5.383-1.739-.713-3.557-.334-5.369-.434v10.82ZM48.794 11.957h-8.421V7.691h15.224c0 .805-.04 1.613.012 2.416.048.74-.177 1.312-.645 1.893-2.532 3.15-5.024 6.333-7.529 9.506-.134.17-.262.345-.471.622h8.734v4.347H40.327c0-1.188-.018-2.375.016-3.56.006-.204.21-.416.355-.6 2.57-3.272 5.145-6.541 7.717-9.813.107-.136.199-.285.379-.546Zm8.264.005c0-1.346 0-2.58.002-3.814 0-.136.023-.272.039-.448h15.177c0 1.146.012 2.275-.014 3.402-.004.17-.171.349-.291.501-2.614 3.317-5.231 6.63-7.846 9.945-.13.163-.245.338-.446.618h.762c2.47 0 4.94.01 7.411-.009.46-.003.63.11.617.599-.033 1.23-.01 2.462-.01 3.73H57.072c0-1.228-.01-2.437.013-3.644.002-.152.161-.31.27-.448 2.509-3.195 5.021-6.388 7.533-9.581.192-.244.38-.492.657-.851h-8.488Z"
+    />
+  </svg>,
+]
