+++ packages/react/src/LogoSuite/LogoSuite.stories.tsx
@@ -1,16 +1,32 @@
 import React from 'react'
 
-import {Meta} from '@storybook/react'
-import {LogoSuite} from './LogoSuite'
+import type {Meta, StoryFn} from '@storybook/react'
+import {LogoSuite, type LogoSuiteHeadingProps, type LogoSuiteLogoBarProps, type LogoSuiteProps} from './LogoSuite'
+import {logos} from './LogoSuite.fixtures'
 
-export default {
+type StoryProps = Required<
+  Pick<LogoSuiteProps, 'align' | 'hasDivider'> &
+    Pick<LogoSuiteHeadingProps, 'visuallyHidden'> &
+    Pick<LogoSuiteLogoBarProps, 'marquee' | 'marqueeSpeed' | 'variant'>
+> & {
+  logoCount: number
+  heading: string
+  description: string
+}
+
+const meta: Meta<StoryProps> = {
   title: 'Components/LogoSuite',
   component: LogoSuite,
   args: {
     align: 'center',
-    variant: 'muted',
-    marqueeSpeed: 'default',
-    hasMarquee: false,
+    hasDivider: true,
+    visuallyHidden: false,
+    variant: 'default',
+    marquee: false,
+    marqueeSpeed: 'normal',
+    logoCount: 5,
+    heading: 'Heading',
+    description: 'Body text maximus ligula felis, non egestas dolor rutrum vel.',
   },
   argTypes: {
     hasDivider: {
@@ -21,12 +37,14 @@ export default {
         category: 'LogoSuite',
       },
     },
-    hasMarquee: {
+    logoCount: {
       control: {
-        type: 'boolean',
+        type: 'number',
+        min: 3,
+        max: 12,
       },
       table: {
-        category: 'LogoSuite.Logobar',
+        category: 'LogoSuite',
       },
     },
     align: {
@@ -34,13 +52,6 @@ export default {
       options: ['start', 'center', 'justify'],
       table: {
         category: 'LogoSuite',
-      },
-    },
-    variant: {
-      control: 'radio',
-      options: ['muted', 'emphasis'],
-      table: {
-        category: 'LogoSuite.Logobar',
       },
     },
     visuallyHidden: {
@@ -51,83 +62,67 @@ export default {
         category: 'LogoSuite.Heading',
       },
     },
+    heading: {
+      control: {
+        type: 'text',
+      },
+      table: {
+        category: 'LogoSuite.Heading',
+      },
+    },
+    description: {
+      control: {
+        type: 'text',
+      },
+      table: {
+        category: 'LogoSuite.Description',
+      },
+    },
+
+    marquee: {
+      control: {
+        type: 'boolean',
+      },
+      table: {
+        category: 'LogoSuite.Logobar',
+      },
+    },
+    variant: {
+      control: 'radio',
+      options: ['default', 'muted', 'emphasis'],
+      table: {
+        category: 'LogoSuite.Logobar',
+      },
+    },
+
     marqueeSpeed: {
       control: 'radio',
-      options: ['slow', 'default', 'idle'],
+      options: ['slow', 'normal', 'idle'],
       table: {
         category: 'LogoSuite.Logobar',
       },
     },
   },
-} as Meta<typeof LogoSuite>
+}
 
-const logos = [
-  <svg key={1} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 115 43" fill="none" focusable="false">
-    <title>Uber</title>
-    <g>
-      <path
-        fill="#24292F"
-        fillRule="evenodd"
-        d="M8.743 33.353c-.847-.902-1.51-1.989-1.99-3.26-.478-1.27-.718-2.68-.718-4.227V1.996H.233v24.202c0 2.284.378 4.356 1.133 6.216.755 1.86 1.787 3.453 3.095 4.78a13.519 13.519 0 0 0 4.641 3.066c1.787.719 3.711 1.078 5.774 1.078 2.027 0 3.914-.368 5.664-1.105a13.322 13.322 0 0 0 4.56-3.15v3.592h5.746V1.995h-5.802v23.871c0 1.547-.24 2.947-.718 4.2-.48 1.252-1.152 2.33-2.018 3.232a9.176 9.176 0 0 1-3.038 2.1 9.368 9.368 0 0 1-3.73.746c-1.363 0-2.625-.24-3.786-.718a8.56 8.56 0 0 1-3.011-2.073Zm32.104-17.267V1.996h-5.581v38.679h5.526v-3.537a14.678 14.678 0 0 0 4.503 3.067c1.713.755 3.564 1.133 5.553 1.133 2.063 0 3.997-.387 5.802-1.16 1.805-.774 3.37-1.824 4.697-3.15a14.906 14.906 0 0 0 3.15-4.67c.773-1.786 1.16-3.692 1.16-5.718s-.387-3.942-1.16-5.747c-.774-1.805-1.824-3.37-3.15-4.697-1.326-1.326-2.892-2.376-4.697-3.15-1.805-.773-3.739-1.16-5.802-1.16-1.989 0-3.83.378-5.525 1.133a14.78 14.78 0 0 0-4.476 3.067Zm2.707 17.46a9.611 9.611 0 0 1-2.072-3.094c-.497-1.178-.746-2.45-.746-3.812 0-1.363.258-2.643.774-3.84a9.913 9.913 0 0 1 2.1-3.123 9.621 9.621 0 0 1 3.094-2.072 9.567 9.567 0 0 1 3.758-.746 9.57 9.57 0 0 1 3.757.746 9.303 9.303 0 0 1 3.067 2.072 10.154 10.154 0 0 1 2.072 3.122 9.603 9.603 0 0 1 .773 3.84c0 1.364-.258 2.635-.773 3.813a10.247 10.247 0 0 1-2.072 3.095 9.666 9.666 0 0 1-3.067 2.1 9.272 9.272 0 0 1-3.758.773c-1.362 0-2.633-.258-3.812-.773a9.999 9.999 0 0 1-3.095-2.1Zm25.252-12.598a15.03 15.03 0 0 1 3.04-4.669 14.12 14.12 0 0 1 4.586-3.177c1.768-.774 3.683-1.16 5.746-1.16 2.026 0 3.905.368 5.637 1.105 1.73.736 3.223 1.759 4.475 3.066 1.253 1.308 2.23 2.855 2.929 4.642.7 1.787 1.05 3.73 1.05 5.83v1.823H73.337a9.58 9.58 0 0 0 3.177 5.719 9.47 9.47 0 0 0 2.819 1.685 9.204 9.204 0 0 0 3.343.608c3.352 0 6.06-1.381 8.123-4.144l4.033 2.984a15.007 15.007 0 0 1-5.194 4.42c-2.063 1.069-4.384 1.603-6.962 1.603-2.1 0-4.061-.369-5.885-1.105-1.823-.737-3.407-1.769-4.752-3.095a14.68 14.68 0 0 1-3.177-4.669c-.774-1.787-1.16-3.73-1.16-5.83-.001-1.989.368-3.867 1.104-5.636Zm7.764-2.182c-1.566 1.307-2.606 3.048-3.122 5.221h17.185c-.48-2.173-1.501-3.914-3.067-5.221-1.566-1.308-3.398-1.962-5.498-1.962s-3.933.654-5.498 1.962Zm28.374 5.884c0-2.21.589-3.96 1.768-5.25 1.179-1.288 2.689-1.933 4.531-1.933h2.321v-5.194h-1.934c-1.547 0-2.883.331-4.006.994a7.853 7.853 0 0 0-2.736 2.708v-3.481h-5.525v28.18h5.581V24.65Z"
-        clipRule="evenodd"
-      />
-    </g>
-    <defs>
-      <clipPath>
-        <path fill="#fff" d="M.233.452h114.75v42.43H.233z" />
-      </clipPath>
-    </defs>
-  </svg>,
-  <svg key={2} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 165 41" fill="none">
-    <title>Vercel</title>
-    <path
-      fill="#24292F"
-      d="M81.939 11.46c-6.33 0-10.9 4.155-10.9 10.374 0 6.22 5.138 10.376 11.468 10.376 3.84 0 7.194-1.521 9.304-4.074l-4.382-2.553c-1.163 1.276-2.92 2.01-4.895 2.01-2.76 0-5.085-1.44-5.95-3.749h16.093a9.74 9.74 0 0 0 .19-2.01c-.028-6.246-4.572-10.375-10.928-10.375Zm-5.437 8.337c.73-2.308 2.678-3.748 5.437-3.748 2.759 0 4.706 1.44 5.436 3.748H76.502Zm67.242-8.338c-6.329 0-10.901 4.156-10.901 10.375 0 6.22 5.14 10.376 11.469 10.376 3.841 0 7.195-1.521 9.304-4.074l-4.381-2.553c-1.163 1.276-2.922 2.01-4.896 2.01-2.759 0-5.085-1.44-5.951-3.749h16.094a9.762 9.762 0 0 0 .189-2.01c0-6.246-4.571-10.375-10.927-10.375Zm-5.41 8.338c.731-2.308 2.678-3.748 5.437-3.748 2.759 0 4.706 1.44 5.437 3.748h-10.874Zm-22.396 2.037c0 3.45 2.245 5.759 5.735 5.759 2.353 0 4.138-1.087 5.058-2.825l4.408 2.553c-1.812 3.042-5.247 4.889-9.466 4.889-6.357 0-10.901-4.156-10.901-10.376s4.571-10.375 10.901-10.375c4.219 0 7.627 1.847 9.466 4.89l-4.408 2.552c-.92-1.765-2.678-2.825-5.058-2.825-3.49-.027-5.735 2.282-5.735 5.758Zm47.335-16.73v26.508h-5.167V5.104h5.167ZM22.189 2.224 43.395 39.11H.983L22.19 2.224Zm53.042 2.88L59.326 32.78 43.422 5.104H49.4l9.953 17.3 9.954-17.3h5.924Zm33.81 6.926v5.595a6.684 6.684 0 0 0-1.839-.272c-3.327 0-5.734 2.309-5.734 5.758v8.528h-5.167V12.03h5.167v5.297c0-2.934 3.381-5.296 7.573-5.296Z"
-    />
-  </svg>,
-  <svg key={3} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 174 50" width="174" height="50" fill="none">
-    <title>Shopify</title>
-    <path
-      fill="#24292F"
-      d="M60.128 27.374c-1.469-.812-2.242-1.469-2.242-2.397 0-1.198 1.044-1.933 2.706-1.933 1.933 0 3.634.812 3.634.812l1.353-4.137s-1.237-.966-4.91-.966c-5.103 0-8.66 2.938-8.66 7.036 0 2.32 1.663 4.098 3.866 5.374 1.779 1.005 2.397 1.74 2.397 2.783 0 1.122-.889 2.01-2.551 2.01-2.474 0-4.794-1.275-4.794-1.275l-1.43 4.137s2.164 1.43 5.76 1.43c5.258 0 9.008-2.59 9.008-7.23.077-2.551-1.856-4.33-4.137-5.644ZM81.082 18.676c-2.59 0-4.6 1.237-6.186 3.093l-.077-.04 2.242-11.752h-5.876l-5.683 29.885h5.838l1.933-10.207c.773-3.866 2.744-6.224 4.6-6.224 1.315 0 1.817.889 1.817 2.165 0 .812-.077 1.778-.27 2.59L77.216 39.9h5.838l2.28-12.062c.271-1.276.426-2.783.426-3.827.077-3.364-1.624-5.335-4.678-5.335ZM99.06 18.676c-7.037 0-11.715 6.34-11.715 13.415 0 4.523 2.784 8.196 8.041 8.196 6.921 0 11.56-6.186 11.56-13.415.038-4.214-2.397-8.196-7.887-8.196Zm-2.862 17.088c-2.01 0-2.822-1.701-2.822-3.828 0-3.363 1.74-8.814 4.91-8.814 2.088 0 2.745 1.778 2.745 3.518 0 3.634-1.74 9.124-4.833 9.124ZM121.946 18.676c-3.943 0-6.186 3.48-6.186 3.48h-.077l.348-3.132h-5.181c-.27 2.126-.734 5.335-1.198 7.77l-4.059 21.341h5.837l1.624-8.66h.116s1.199.773 3.441.773c6.881 0 11.366-7.036 11.366-14.15 0-3.904-1.74-7.422-6.031-7.422Zm-5.567 17.204c-1.508 0-2.397-.85-2.397-.85l.966-5.452c.696-3.634 2.591-6.07 4.601-6.07 1.778 0 2.32 1.663 2.32 3.21 0 3.75-2.243 9.162-5.49 9.162ZM136.328 10.286c-1.856 0-3.364 1.47-3.364 3.402 0 1.74 1.083 2.938 2.745 2.938h.077c1.817 0 3.402-1.237 3.441-3.402.039-1.74-1.121-2.938-2.899-2.938ZM128.132 39.862h5.876l3.982-20.722h-5.915l-3.943 20.722ZM152.836 19.1h-4.06l.194-.966c.347-2.01 1.507-3.788 3.479-3.788 1.044 0 1.856.309 1.856.309l1.159-4.562s-1.005-.503-3.17-.503c-2.087 0-4.136.58-5.721 1.933-2.011 1.701-2.939 4.137-3.402 6.611l-.155.967h-2.706l-.851 4.407h2.706l-3.092 16.353h5.837l3.093-16.353h4.021l.812-4.407ZM166.908 19.14s-3.673 9.2-5.297 14.227h-.077c-.116-1.624-1.43-14.227-1.43-14.227h-6.147l3.518 19.02c.077.426.038.696-.116.967-.696 1.314-1.817 2.59-3.17 3.518-1.083.812-2.32 1.315-3.286 1.662l1.623 4.949c1.199-.27 3.634-1.237 5.722-3.17 2.668-2.513 5.142-6.34 7.655-11.598l7.152-15.348h-6.147ZM29.354 5.686s-.54.154-1.43.425a11.31 11.31 0 0 0-.696-1.701c-1.005-1.933-2.513-2.977-4.291-2.977-.116 0-.232 0-.387.039-.038-.078-.116-.116-.154-.194C21.622.428 20.617.041 19.419.08c-2.32.077-4.64 1.74-6.495 4.716-1.315 2.088-2.32 4.717-2.59 6.766-2.668.812-4.524 1.392-4.562 1.43-1.353.426-1.392.464-1.547 1.74C4.11 15.699.552 42.915.552 42.915L29.78 47.98V5.647c-.232 0-.348.039-.426.039ZM22.59 7.773c-1.546.464-3.248 1.005-4.91 1.508.464-1.817 1.392-3.634 2.474-4.833.425-.425 1.005-.927 1.663-1.237.657 1.392.812 3.286.773 4.562Zm-3.17-6.108c.541 0 1.005.116 1.392.348-.62.31-1.238.812-1.817 1.392-1.47 1.585-2.59 4.059-3.055 6.417-1.392.425-2.783.85-4.059 1.237.85-3.711 3.982-9.317 7.539-9.394Zm-4.485 21.263c.155 2.474 6.688 3.015 7.075 8.853.27 4.6-2.436 7.732-6.34 7.964-4.717.31-7.307-2.474-7.307-2.474l1.005-4.253s2.59 1.972 4.678 1.817c1.353-.077 1.856-1.198 1.817-1.971-.193-3.248-5.528-3.054-5.876-8.39-.31-4.484 2.628-9.007 9.123-9.433 2.513-.154 3.79.464 3.79.464l-1.47 5.567s-1.662-.773-3.634-.618c-2.86.193-2.9 2.01-2.86 2.474Zm9.201-15.619c0-1.16-.154-2.822-.696-4.214 1.779.348 2.63 2.32 3.016 3.518-.696.194-1.47.426-2.32.696ZM30.669 47.903l12.14-3.016S37.588 9.59 37.55 9.358a.45.45 0 0 0-.426-.386c-.193 0-3.595-.078-3.595-.078s-2.088-2.01-2.861-2.783v41.792Z"
-    />
-  </svg>,
-  <svg key={4} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 175 45" fill="none">
-    <title>Pinterest</title>
-    <path
-      fill="#24292F"
-      fillRule="evenodd"
-      d="M.088 16.766c0-8.45 6.12-16.23 17.716-16.23 9.295 0 16.521 6.644 16.55 15.53 0 9.267-5.827 16.697-13.927 16.697-2.71 0-5.274-1.428-6.149-3.089 0 0-1.34 5.129-1.66 6.381-.904 3.468-4.109 8.276-4.371 8.625a.245.245 0 0 1-.204.117c-.146 0-.292-.117-.292-.262-.087-.496-1.02-6.265-.204-9.82.496-2.098 3.147-13.345 3.147-13.345a9.424 9.424 0 0 1-.815-3.992c0-3.73 2.156-6.527 4.866-6.527 2.272 0 3.38 1.72 3.38 3.788 0 1.424-.558 3.284-1.15 5.254-.363 1.213-.74 2.47-1.007 3.691-.64 2.652 1.34 4.837 3.963 4.837 4.75 0 8.421-5.011 8.421-12.267 0-6.41-4.604-10.898-11.189-10.898-7.634 0-12.121 5.712-12.121 11.627 0 2.302.874 4.778 1.981 6.119.175.233.262.524.175.786-.106.44-.283 1.137-.44 1.76-.146.577-.276 1.089-.318 1.27-.116.496-.379.583-.903.35C2.187 25.595.088 20.7.088 16.766Zm91.377 4.837c1.865 0 2.506 1.137 2.506 2.477 0 2.273-1.748 3.642-4.517 3.642a12.9 12.9 0 0 1-3.176-.495c.35-2.098 1.69-5.624 5.187-5.624Zm39.773 0c1.69 0 2.506 1.137 2.506 2.477 0 2.273-1.748 3.642-4.516 3.642a12.901 12.901 0 0 1-3.176-.495c.35-2.098 1.865-5.624 5.186-5.624Zm37.705.175h4.225c.583 0 .933-.233 1.049-.787.074-.286.164-.65.254-1.012v-.002c.157-.635.313-1.27.387-1.492.117-.495-.146-.845-.699-.845h-4.196c.051-.207.128-.496.222-.845l.001-.002c.448-1.663 1.263-4.692 1.263-6.787 0-5.07-4.312-9.382-10.431-9.382-6.906 0-11.393 4.633-11.393 11.13a10.809 10.809 0 0 0 1.602 5.508c-.379-.03-.757-.03-1.136-.03-5.245 0-8.538 2.856-8.538 6.994 0 3.875 3.06 5.39 5.944 6.41.393.138.78.265 1.155.387 2.077.679 3.77 1.232 3.77 2.935 0 1.34-1.282 2.127-3.846 2.127-3.002 0-4.662-1.428-5.1-1.953a2.08 2.08 0 0 0 .7-1.66c0-1.253-.874-2.244-2.506-2.244-1.37 0-2.535 1.107-2.827 2.593-2.214 1.807-4.603 2.972-8.012 2.972-3.38 0-5.1-1.748-5.1-5.477.787.233 2.856.524 4.196.524 5.682 0 9.558-2.593 9.558-7.78 0-2.797-2.448-5.711-7.285-5.711h-.029c-6.352 0-9.936 4.021-11.481 8.48-.218-.06-.437-.11-.655-.161-.219-.051-.437-.102-.656-.16a9.626 9.626 0 0 0 .612-3.38c0-2.186-1.369-4.779-4.982-4.779-2.535 0-4.925 1.66-6.586 4.138.186-.742.347-1.364.468-1.83.145-.558.232-.89.232-.939.116-.437-.059-.903-.729-.903h-3.729c-.525 0-.904.175-1.049.787-.088.408-1.37 5.449-2.623 10.315-.787 3.06-4.837 6.847-9.936 6.847-3.35 0-5.1-1.865-5.1-5.594.875.35 2.827.64 4.197.64 5.915 0 9.586-2.592 9.586-7.78 0-2.796-2.447-5.71-7.284-5.71h-.03c-7.167 0-11.538 5.565-12.296 11.888-.32 2.623-2.535 6.236-4.924 6.236-1.166 0-1.836-.729-1.836-1.982 0-.851.473-2.665 1.087-5.027.164-.629.338-1.296.516-1.995.188-.715.425-1.673.694-2.764l.002-.006c.147-.597.305-1.234.47-1.892h3.904c.583 0 .903-.233 1.049-.787.175-.786.524-2.156.64-2.506.117-.495-.145-.845-.698-.845h-3.847s1.749-7.022 1.807-7.226c.175-.728-.437-1.136-1.05-.99 0 0-2.971.582-3.525.699-.553.116-.99.437-1.194 1.224l-1.574 6.322h-3.06c-.582 0-.932.234-1.048.787-.075.286-.164.65-.254 1.013-.157.635-.313 1.27-.387 1.493-.117.495.145.845.699.845h3.001c-.029.088-1.078 4.138-1.952 7.78-.408 1.748-1.778 5.857-3.992 5.857-1.34 0-1.894-.67-1.894-2.127 0-1.102.528-3.006 1.071-4.967.577-2.084 1.173-4.231 1.173-5.552 0-3.438-1.836-5.449-5.624-5.449-2.39 0-4.866 1.544-5.915 2.885 0 0 .32-1.108.437-1.545a.676.676 0 0 0-.437-.845c-.087-.029-.204-.029-.291-.029h-3.672c-.787 0-.99.437-1.107.874-.032.095-.429 1.669-.99 3.89-.477 1.892-1.072 4.253-1.662 6.57-.874 3.468-3 6.353-5.274 6.353-1.165 0-1.69-.728-1.69-1.981 0-.945.537-3.098 1.22-5.831.095-.387.195-.784.296-1.192a1023.18 1023.18 0 0 1 1.396-5.433c.326-1.256.544-2.095.585-2.288.117-.495-.087-.904-.729-.904h-3.73c-.67 0-.932.35-1.048.845 0 0-1.05 3.963-2.156 8.392l-.223.873c-.745 2.91-1.467 5.736-1.467 7.14 0 2.74 1.223 4.808 4.603 4.808 2.594 0 4.663-1.311 6.236-3.001l-.14.546c-.151.587-.245.953-.268.998-.146.525.03.99.641.99h3.817c.67 0 .903-.261 1.05-.844.145-.583 2.971-11.743 2.971-11.743.758-3.001 2.623-5.012 5.245-5.012 1.253 0 2.302.816 2.185 2.419-.057.691-.42 2.097-.838 3.716-.638 2.476-1.405 5.45-1.405 7.123 0 2.098.786 4.808 4.72 4.808 2.68 0 4.691-1.282 6.12-2.972.524 1.777 1.922 2.972 4.486 2.972 4.255 0 6.819-2.535 8.334-5.07 1.253 2.943 4.196 5.07 8.48 5.07a13.727 13.727 0 0 0 10.168-4.37c-.208.803-.382 1.488-.512 1.995-.143.559-.23.901-.245.947-.146.525.087.962.728.962h3.759a.981.981 0 0 0 1.02-.816c.025-.112.141-.562.325-1.275l.001-.004v-.002c.245-.95.611-2.369 1.043-4.08 1.486-5.711 3.846-11.656 7.401-11.656 1.253 0 1.778.7 1.778 1.807 0 .408-.088.845-.292 1.224-1.719-.35-3.117.495-3.117 2.33 0 1.195 1.282 2.303 2.972 2.303 1.311 0 2.389-.32 3.234-.904.306.069.626.137.954.206h.001c.513.11 1.047.224 1.58.348a17.73 17.73 0 0 0-.204 2.506c0 4.604 3.438 8.508 9.324 8.508 4.604 0 7.576-1.544 10.344-3.846 1.311 2.156 4.633 3.875 8.975 3.875 5.944 0 9.353-3.088 9.353-7.197 0-3.7-3.059-5.07-6.177-6.206-.311-.113-.616-.22-.911-.323-2.138-.75-3.78-1.324-3.78-2.912 0-1.486 1.165-2.156 3.147-2.156.932 0 1.893.146 2.797.437.524 1.282 1.573 2.71 3.38 2.71 1.544 0 2.214-1.224 2.214-2.33 0-1.288-.87-2.11-1.906-3.09-1.52-1.437-3.397-3.21-3.397-7.284 0-4.254 2.419-7.75 7.401-7.75 3.468 0 5.74 2.185 5.74 5.74 0 2.913-1.777 8.129-1.777 8.129h-3.001c-.583 0-.933.233-1.049.787-.074.286-.164.65-.254 1.012-.156.636-.313 1.271-.387 1.494-.117.495.145.845.699.845h2.943s-3.03 11.218-3.03 13.257c0 2.74 1.544 4.808 4.924 4.808 4.779 0 7.78-3.089 9.062-7.605a.484.484 0 0 0-.233-.612c-.058-.029-.146-.029-.204-.029h-1.836a.686.686 0 0 0-.67.554c-.408 1.602-1.428 3.292-3.438 3.292-1.166 0-1.836-.728-1.836-1.981 0-.953.546-3.021 1.297-5.863l.306-1.16c.437-1.748 1.165-4.778 1.165-4.778Zm-124.594-9.85c-.146 1.982-1.777 3.556-3.671 3.556s-3.322-1.603-3.206-3.555c.146-1.982 1.778-3.555 3.672-3.555 1.894 0 3.322 1.603 3.205 3.555Z"
-      clipRule="evenodd"
-    />
-  </svg>,
-  <svg key={5} role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 134 40" width="134" height="40" fill="none">
-    <title>Twilio</title>
-    <path
-      fill="#57606A"
-      fillRule="evenodd"
-      d="M93.765 7.84h-7.04c-.16 0-.32.16-.32.32v4.08c0 .16.16.32.32.32h7.04c.16 0 .32-.16.32-.32V8.16c.08-.16-.08-.32-.32-.32Zm-.08 6.08h-13.6c-.16 0-.4.16-.48.32l-1.76 6.64-.08.32-2.08-7.12c-.08-.16-.24-.32-.48-.32h-5.28c-.16 0-.4.16-.48.32l-2 6.64-.08.32-.08-.16-.8-3.28-.88-3.28c-.08-.16-.24-.32-.48-.32h-10.64V8.24c0-.16-.24-.4-.48-.32L47.285 10c-.24.08-.32.16-.32.32v3.6h-1.76c-.16 0-.32.16-.32.32v5.12c0 .16.16.32.32.32h1.76v6.24c0 4.4 2.4 6.48 6.8 6.48 1.84 0 3.6-.4 4.8-1.04v-5.28c0-.24-.24-.32-.4-.32-.72.32-1.28.4-1.92.4-1.2 0-1.84-.48-1.84-1.84v-4.56h3.84c.16 0 .32-.16.32-.32v-4.32l5.04 16.72c.08.16.24.32.48.32h5.6c.16 0 .4-.16.48-.32l2.4-7.44 1.2 3.84 1.12 3.52c.08.16.24.32.48.32h5.6c.16 0 .4-.16.48-.32l5.12-16.8v16.8c0 .16.16.32.32.32h6.8c.16 0 .32-.16.32-.32V14.32c0-.24-.16-.4-.32-.4Zm8.96-6.08h-6.8c-.16 0-.32.16-.32.32v23.6c0 .16.16.32.32.32h6.8c.16 0 .32-.16.32-.32V8.16c0-.16-.16-.32-.32-.32Zm9.04 0h-7.04c-.16 0-.32.16-.32.32v4.08c0 .16.16.32.32.32h7.04c.16 0 .32-.16.32-.32V8.16c.08-.16-.08-.32-.32-.32Zm-.08 6.08h-6.8c-.16 0-.32.16-.32.32v17.52c0 .16.16.32.32.32h6.8c.16 0 .32-.16.32-.32V14.32c0-.24-.16-.4-.32-.4Zm11.28-.4c-6 0-10.24 4.4-10.24 9.52v.08c0 5.12 4.24 9.44 10.16 9.44 6 0 10.24-4.4 10.24-9.52v-.08c0-5.12-4.24-9.44-10.16-9.44Zm2.88 9.52c0 1.76-1.2 3.12-2.88 3.12s-2.96-1.44-2.96-3.2v-.08c0-1.76 1.2-3.12 2.88-3.12s2.96 1.44 2.96 3.28ZM20.085 0c-11.04 0-20 8.96-20 20s8.96 20 20 20 20-8.96 20-20-8.96-20-20-20Zm0 34.72c-8.16 0-14.72-6.56-14.72-14.72s6.56-14.72 14.72-14.72 14.72 6.56 14.72 14.72-6.56 14.72-14.72 14.72Zm4.96-15.52a4.16 4.16 0 1 0 0-8.32 4.16 4.16 0 0 0 0 8.32Zm4.16 5.76a4.16 4.16 0 1 1-8.32 0 4.16 4.16 0 0 1 8.32 0Zm-14.08 4.16a4.16 4.16 0 1 0 0-8.32 4.16 4.16 0 0 0 0 8.32Zm4.16-14.08a4.16 4.16 0 1 1-8.32 0 4.16 4.16 0 0 1 8.32 0Z"
-      clipRule="evenodd"
-    />
-  </svg>,
-]
+export default meta
 
-export const Default = () => (
-  <LogoSuite>
-    <LogoSuite.Heading>Heading</LogoSuite.Heading>
-    <LogoSuite.Description>Body text maximus ligula felis, non egestas dolor rutrum vel.</LogoSuite.Description>
-    <LogoSuite.Logobar>
-      <React.Fragment>{logos.slice(0, 5)}</React.Fragment>
-    </LogoSuite.Logobar>
-  </LogoSuite>
-)
-
-export const Playground = args => (
-  <LogoSuite {...args}>
-    <LogoSuite.Heading visuallyHidden={args.visuallyHidden}>Heading</LogoSuite.Heading>
-    <LogoSuite.Description>Body text maximus ligula felis, non egestas dolor rutrum vel.</LogoSuite.Description>
-    <LogoSuite.Logobar variant={args.variant} marquee={args.hasMarquee} marqueeSpeed={args.marqueeSpeed}>
-      <React.Fragment>{logos.slice(0, 5)}</React.Fragment>
+export const Playground: StoryFn<StoryProps> = ({
+  hasDivider,
+  logoCount,
+  align,
+  visuallyHidden,
+  marquee,
+  marqueeSpeed,
+  variant,
+  heading,
+  description,
+}) => (
+  <LogoSuite align={align} hasDivider={hasDivider}>
+    <LogoSuite.Heading visuallyHidden={visuallyHidden}>{heading}</LogoSuite.Heading>
+    <LogoSuite.Description>{description}</LogoSuite.Description>
+    <LogoSuite.Logobar variant={variant} marquee={marquee} marqueeSpeed={marqueeSpeed}>
+      {logos.slice(0, logoCount)}
     </LogoSuite.Logobar>
   </LogoSuite>
 )
